### 2024-09-14 #212 @adtzlr
FElupe submission
Submitting Author: Andreas  (@adtzlr)
Package Name: FElupe
One-Line Description of Package: Finite element analysis for continuum mechanics of solid bodies.
Repository Link: https://github.com/adtzlr/felupe
Version submitted: v9.0.0
EiC: TBD
Version accepted: TBD

---

-
FElupe is a Python finite element analysis package focusing on the formulation and numerical solution of nonlinear problems in continuum mechanics of solid bodies. Easy-to-learn classes are provided to simulate the nonlinear deformation of hyperelastic solid bodies, see [Getting Started](https://github.com/adtzlr/felupe?tab=readme-ov-file#getting-started). The constitutive material formulation of a hyperelastic solid body may be defined by its strain energy density function only (gradient and hessian are carried out by optional [automatic differentiation](https://github.com/adtzlr/tensortrax)). Strain energy functions for [selected hyperelastic models](https://felupe.readthedocs.io/en/latest/felupe/constitution/hyperelasticity.html) are included. FElupe has only few dependencies, is a pure Python package but is also efficient enough for [rubber-like structures](https://felupe.readthedocs.io/en/latest/examples/ex06_rubber-metal-spring.html#sphx-glr-examples-ex06-rubber-metal-spring-py). Several [extension packages]((https://github.com/adtzlr/felupe?tab=readme-ov-file#extension-packages)) exist, e.g. to use a different automatic differentation backend or an interactive plot window.

[![grafik](https://github.com/user-attachments/assets/5584ca5e-1677-40f9-8f9e-d4c04f03ef1d)](https://github.com/adtzlr/felupe)

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

FElupe is great for teaching, scientific research as well as for mid-sized industry-problems related to the deformation of solid bodies. It has already been used in scientific articles which are listed in the [README](https://github.com/adtzlr/felupe?tab=readme-ov-file#scientific-publications).

Yes, there is [scikit-fem](https://github.com/kinnala/scikit-fem), [FEniCS](https://fenicsproject.org/), [GetFEM](https://getfem.org) and probably some others. While [scikit-fem](https://github.com/kinnala/scikit-fem) is definitely lightweight and easy-to-install, I found out that it is typically too slow for running hyperelastic simulations, e.g. the deformation of a rubber-like solid (see also my initial questions / posts in https://github.com/kinnala/scikit-fem/issues/616). [FEniCS](https://fenicsproject.org/) on the other hand has a great community but the install-entry-barrier is much higher compared to a pure PyPI Python package. The same also applies for [GetFEM](https://getfem.org), because both projects provide Python packages for their compiled code. I was looking for a more lightweight Python package but also efficient enough for [simulating typical rubber-like structures](https://felupe.readthedocs.io/en/latest/examples/ex05_rubber-metal-bushing.html#sphx-glr-examples-ex05-rubber-metal-bushing-py). Hence, I started working on FElupe in 2021.

Instead of assembling general weak forms like [scikit-fem](https://github.com/kinnala/scikit-fem) and [FEniCS](https://fenicsproject.org/), FElupe follows a more Abaqus-like UMAT (user material) approach for constitutive material formulations of [solid bodies](https://felupe.readthedocs.io/en/latest/felupe/mechanics.html#felupe.SolidBody), see e.g. [ConstitutiveMaterial](https://felupe.readthedocs.io/en/latest/felupe/constitution.html#felupe.ConstitutiveMaterial) or [Hyperelastic](https://felupe.readthedocs.io/en/latest/felupe/constitution/hyperelasticity.html#felupe.Hyperelastic).

#211

     **Note**: *Will be added soon - not present in the submitted v9.0.0*.
- [x] The package is deposited in a long-term repository with the DOI: https://zenodo.org/doi/10.5281/zenodo.4817406
