### 2024-09-21 727dbd0
Update _scene.py
