### 2024-09-21 ff9ea3e
enhance docstrings of view-classes
