### 2024-09-21 c2fd524
Add more examples in the docstrings of lower-level classes
