### 2024-09-21 #212 @cmarmo
Hi @adtzlr! Thank you for submitting your package for pyOpenSci review.
Sorry to have kept you waiting!

Below are the basic checks that your package needs to pass to begin our review.
If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

---
---

Your package is already in great shape! Congratulation!
While my background is quite far from continuum mechanics of solid bodies, I particularly enjoyed the documentation and the graphical examples.

Here some clarifications about the unchecked boxes:
- your `README.md` is very detailed but it is missing a link to the development instructions: in fact there are no development instructions in the documentation. Do you mind adding some? This will also allows you to link them in the `CONTRIBUTING.md` file.
- Issue templates are not available in the `.github` repository: as your `CONTRIBUTING.md` clearly defines different kind of contributions, then issues, creating the templates will direct the contributor to the right type of feedback.
- It looks like the majority of the links in the `CONTRIBUTING.md` gives a 404 not found error. Do you mind checking them?

Once those items are fixed we probably already have an editor for your submission... :rocket:
