### 2024-09-21 #212 @adtzlr
Thank you very much @cmarmo for the basic checks and your detailed comments on what is missing. I'll enhance the mentioned files and add the missing issue templates.
