### 2024-09-21 adtzlr/felupe#853 @adtzlr
Basic checks for PyOpenSci submission
The following points must be fixed for the submission https://github.com/pyOpenSci/software-submission/issues/212, see https://github.com/pyOpenSci/software-submission/issues/212#issuecomment-2365311571:

- [ ] Add an instruction on how to install the dev-version in `README.md`
- [ ] Add issue templates (according to `CONTRIBUTING.md`)
- [ ] Fix the links in `CONTRIBUTING.md`
