### 2024-09-21 fac55bb
Update CONTRIBUTING.md
fix links and refer also to discussions
