+++ CONTRIBUTING.md
@@ -1,3 +1,4 @@
+<!-- omit in toc -->
 # Contributing to FElupe
 
 First off, thanks for taking the time to contribute! ❤️
@@ -10,7 +11,7 @@ All types of contributions are encouraged and valued. See the [Table of Contents
 > - Refer this project in your project's readme
 > - Mention the project at local meetups and tell your friends/colleagues
 
-
+<!-- omit in toc -->
 ## Table of Contents
 
 - [Code of Conduct](#code-of-conduct)
@@ -19,63 +20,58 @@ All types of contributions are encouraged and valued. See the [Table of Contents
 - [Reporting Bugs](#reporting-bugs)
 - [Suggesting Enhancements](#suggesting-enhancements)
 
-
 ## Code of Conduct
 
 This project and everyone participating in it is governed by the
-[CONTRIBUTING.md Code of Conduct](blob/main/CODE_OF_CONDUCT.md).
+[CONTRIBUTING.md Code of Conduct](https://github.com/adtzlr/felupe/blob/main/CODE_OF_CONDUCT.md).
 By participating, you are expected to uphold this code. Please report unacceptable behavior
-to <>.
-
+to <a.dutzler@gmail.com>.
 
 ## I Have a Question
 
-> If you want to ask a question, we assume that you have read the available [Documentation]().
+> If you want to ask a question, we assume that you have read the available [Documentation](https://felupe.readthedocs.io/).
 
-Before you ask a question, it is best to search for existing [Issues](/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first.
+Before you ask a question, it is best to search for existing [Issues](https://github.com/adtzlr/felupe/issues) and [Discussions](https://github.com/adtzlr/felupe/discussions) that might help you. In case you have found a suitable issue or discussion and still need clarification, you can write your question in this issue or discussion. It is also advisable to search the internet for answers first.
 
 If you then still feel the need to ask a question and need clarification, we recommend the following:
 
-- Open an [Issue](/issues/new).
+- Open a [Discussion](https://github.com/adtzlr/felupe/discussions/new/choose).
 - Provide as much context as you can about what you're running into.
-- Provide project and platform versions (nodejs, npm, etc), depending on what seems relevant.
+- Provide project and platform versions, depending on what seems relevant.
 
 We will then take care of the issue as soon as possible.
-
-
 
 ## I Want To Contribute
 
 > ### Legal Notice
-> When contributing to this project, you must agree that you have authored 100% of the content, that you have the necessary rights to the content and that the content you contribute may be provided under the project [license](blob/main/LICENSE).
+> When contributing to this project, you must agree that you have authored 100% of the content, that you have the necessary rights to the content and that the content you contribute may be provided under the project [license](https://github.com/adtzlr/felupe/blob/main/LICENSE).
 
 ### Reporting Bugs
 
-
+<!-- omit in toc -->
 #### Before Submitting a Bug Report
 
 A good bug report shouldn't leave others needing to chase you up for more information. Therefore, we ask you to investigate carefully, collect information and describe the issue in detail in your report. Please complete the following steps in advance to help us fix any potential bug as fast as possible.
 
 - Make sure that you are using the latest version.
 - Determine if your bug is really a bug and not an error on your side e.g. using incompatible environment components/versions (Make sure that you have read the [documentation](https://felupe.readthedocs.io/). If you are looking for support, you might want to check [this section](#i-have-a-question)).
-- To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the [bug tracker](issues?q=label%3Abug).
[…]
