### 2024-09-21 1b5edef
Simplify the install section in README.md
