### 2024-09-21 60e7384
Fix pyOpenSci basic checks (#854)
* Update CONTRIBUTING.md

fix links and refer also to discussions

* Update CODE_OF_CONDUCT.md

* Simplify the install section in README.md

* Add development install instructions to the docs

* Update README.md
