+++ CONTRIBUTING.md
@@ -80,7 +80,7 @@ Once it's filed:
 
 - The project team will label the issue accordingly.
 - A team member will try to reproduce the issue with your provided steps. If there are no reproduction steps or no obvious way to reproduce the issue, the team will ask you for those steps.
-- If the team is able to reproduce the issue, the issue will be left to be [implemented by someone](#your-first-code-contribution).
+- If the team is able to reproduce the issue, the issue will be left to be implemented by someone.
 
 <!-- omit in toc -->
 ### Suggesting Enhancements
