+++ CONTRIBUTING.md
@@ -69,9 +69,9 @@ A good bug report shouldn't leave others needing to chase you up for more inform
 
 > You must never report security related issues, vulnerabilities or bugs including sensitive information to the issue tracker, or elsewhere in public. Instead sensitive bugs must be sent by email to <a.dutzler@gmail.com>.
 
-We use GitHub issues to track bugs and errors. If you run into an issue with the project:
+We use [GitHub issues](https://github.com/adtzlr/felupe/issues) to track bugs and errors. If you run into an issue with the project:
 
-- Open an [Issue](https://github.com/adtzlr/felupe/issues/new). (Since we can't be sure at this point whether it is a bug or not, we ask you not to talk about a bug yet and not to label the issue.)
+- Open an [Issue](https://github.com/adtzlr/felupe/issues/new/choose).
 - Explain the behavior you would expect and the actual behavior.
 - Please provide as much context as possible and describe the *reproduction steps* that someone else can follow to recreate the issue on their own. This usually includes your code. For good bug reports you should isolate the problem and create a reduced test case.
 - Provide the information you collected in the previous section.
