### 2024-09-22 de2ea6b
Enhance the highlights section
add the base cartesian field
