### 2024-09-23 95ec435
Extract field data as C-contiguous array
this enhances the time spent on assembly
