### 2024-09-23 2fc49e2
Change to default `np.einsum(..., order="C")` in `Field`
