+++ src/felupe/field/_base.py
@@ -308,7 +308,7 @@ class Field:
             shape that the inputs broadcast to. If not provided or None, a freshly-
             allocated array is returned (default is None).
         order : {'C', 'F', 'A', 'K'}, optional
-            Controls the memory layout of the outputs. 'C' means it should be C
+            Controls the memory layout of the output. 'C' means it should be C
             contiguous. 'F' means it should be Fortran contiguous, 'A' means it should
             be 'F' if the inputs are all 'F', 'C' otherwise. 'K' means it should be as
             close to the layout as the inputs as is possible, including arbitrarily
