+++ src/felupe/field/_planestrain.py
@@ -67,7 +67,7 @@ class FieldPlaneStrain(Field):
         # init base Field
         super().__init__(region, dim=dim, values=values, dtype=dtype)
 
-    def _interpolate_2d(self, dtype=None, out=None):
+    def _interpolate_2d(self, dtype=None, out=None, order="C"):
         """Interpolate 2D field values at points and evaluate them at the
         integration points of all cells in the region."""
 
@@ -80,19 +80,21 @@ class FieldPlaneStrain(Field):
             self.region.h,
             dtype=dtype,
             out=out,
+            order=order,
         )
 
-    def interpolate(self, dtype=None, out=None):
+    def interpolate(self, dtype=None, out=None, order="C"):
         # out-argument is not supported
         # if out is not None:
         #     out = out[:2]
 
-        # extend dimension of in-plane 2d-gradient
+        # extend dimension of in-plane 2d-gradient (out-keyword can't be used here)
         return np.pad(
-            self._interpolate_2d(dtype=dtype, out=None), ((0, 1), (0, 0), (0, 0))
+            self._interpolate_2d(dtype=dtype, out=None, order=order),
+            ((0, 1), (0, 0), (0, 0)),
         )
 
-    def _grad_2d(self, sym=False, dtype=None, out=None):
+    def _grad_2d(self, sym=False, dtype=None, out=None, order="C"):
         """In-plane 2D gradient as partial derivative of field values at points
         w.r.t. the undeformed coordinates, evaluated at the integration points
         of all cells in the region. Optionally, the symmetric part of the
@@ -109,6 +111,12 @@ class FieldPlaneStrain(Field):
             A location into which the result is stored. If provided, it must have a
             shape that the inputs broadcast to. If not provided or None, a freshly-
             allocated array is returned (default is None).
+        order : {'C', 'F', 'A', 'K'}, optional
+            Controls the memory layout of the output. 'C' means it should be C
+            contiguous. 'F' means it should be Fortran contiguous, 'A' means it should
+            be 'F' if the inputs are all 'F', 'C' otherwise. 'K' means it should be as
+            close to the layout as the inputs as is possible, including arbitrarily
+            permuted axes. Default is 'C'.
 
         Returns
         -------
@@ -127,6 +135,7 @@ class FieldPlaneStrain(Field):
             self.region.dhdX,
             dtype=dtype,
             out=out,
+            order=order,
         )
 
         if sym:
@@ -134,7 +143,7 @@ class FieldPlaneStrain(Field):
         else:
             return g
 
-    def _hess_2d(self, dtype=None, out=None):
+    def _hess_2d(self, dtype=None, out=None, order="C"):
         r"""In-plane 2D Hessian as second partial derivative of field values w.r.t.
         undeformed coordinates, evaluated at the integration points of all cells in the
         region.
@@ -148,6 +157,12 @@ class FieldPlaneStrain(Field):
             A location into which the result is stored. If provided, it must have a
             shape that the inputs broadcast to. If not provided or None, a freshly-
             allocated array is returned (default is None).
+        order : {'C', 'F', 'A', 'K'}, optional
+            Controls the memory layout of the output. 'C' means it should be C
+            contiguous. 'F' means it should be Fortran contiguous, 'A' means it should
+            be 'F' if the inputs are all 'F', 'C' otherwise. 'K' means it should be as
+            close to the layout as the inputs as is possible, including arbitrarily
+            permuted axes. Default is 'C'.
 
         Returns
         -------
@@ -166,11 +181,12 @@ class FieldPlaneStrain(Field):
             self.region.d2hdXdX,
             dtype=dtype,
             out=out,
+            order=order,
         )
 
         return h
 
[…]
