### 2024-09-23 9a83fdd
Extract C-contiguous arrays by default from fields (#856)
* Extract field data as C-contiguous array

this enhances the time spent on assembly

* Change to default `np.einsum(..., order="C")` in `Field`

* Update _container.py
