+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 - Add the hessian of the element shape functions for a quadratic quad element `QuadraticQuad.hessian()`.
+- Add the `order`-argument to `FieldContainer.extract(order="C")` to return C-contiguous arrays by default.
+
+### Changed
+- Change default `np.einsum(..., order="K")` to `np.einsum(..., order="C")` in the methods of `Field`, `FieldAxisymmetric` and `FieldPlaneStrain`.
 
 ## [9.0.0] - 2024-09-06
 
