### 2024-09-23 c918ce1
Docs: update benchmark results
