### 2024-09-23 9ca89df
Docs: Run the beginner's guide in interactive mode
