### 2024-09-23 7744b8d
Docs: Switch to static plots for some 2d meshes
