### 2024-09-23 8268e9d
Force more static plots in mesh-api examples
hopefulle to reduce read-the-docs build errors
