+++ docs/howto/solvers.rst
@@ -1,14 +1,16 @@
 Use external solvers
 --------------------
 
-FElupe uses SuperLU as direct sparse solver by default because it is shipped with SciPy (and SciPy is already a dependency of FElupe). While it is definitely a good choice for small to mid-sized problems, faster alternatives are easy to install and use. This section demonstrates several possibilities, e.g. a fast direct solver from `PyPardiso <https://github.com/haasad/PyPardisoProject>`_ (``pip install pypardiso``) and the ``minres`` iterative solver from ``SciPy``. Custom solvers may be passed to the evaluation of a job.
+FElupe uses SuperLU as direct sparse solver by default because it is shipped with SciPy (and SciPy is already a dependency of FElupe). While it is definitely a good choice for small to mid-sized problems, faster alternatives are easy to install and use. This section demonstrates several possibilities, e.g. a fast direct solver from `PyPardiso <https://github.com/haasad/PyPardisoProject>`_ (``pip install pypardiso``) and the :func:`~scipy.sparse.linalg.minres` iterative solver from ``SciPy``. Custom solvers may be passed to the evaluation of a :class:`~felupe.Job`.
 
 ..  code-block:: python
 
     import felupe as fem
+    
+    from scipy.sparse.linalg import spsolve  # (default)
 
     job = fem.Job(steps)
-    job.evaluate(solver=solver) # function `x = solver(A, b)`
+    job.evaluate(solver=spsolve) # function `x = spsolve(A, b)`
 
 Solvers from SciPy Sparse:
 
@@ -17,7 +19,7 @@ Solvers from SciPy Sparse:
    .. code-block:: python
       
       # the default solver
-      from scipy.sparse.linalg import spsolve as solver
+      from scipy.sparse.linalg import spsolve
 
 .. tab:: SciPy Sparse (iterative)
 
@@ -30,7 +32,7 @@ Solvers from SciPy Sparse:
        import scipy.sparse.linalg as spla
 
        def solver(A, b):
-           "Wrapper function for iterative solvers from scipy.sparse."
+           "Wrapper function for iterative solvers from scipy.sparse.linalg."
            
            return spla.minres(A, b)[0]
 
@@ -46,7 +48,7 @@ Solvers from external packages:
 
    ..  code-block:: python
       
-       from pypardiso import spsolve as solver
+       from pypardiso import spsolve
 
 .. tab:: PyPardiso (direct, symmetric)
 
@@ -61,7 +63,7 @@ Solvers from external packages:
        from pypardiso import PyPardisoSolver
        from scipy.sparse import triu
 
-       def solver(A, b):
+       def spsolve(A, b):
            # mtype = 1: real and structurally symmetric, supernode pivoting
            # mtype = 2: real and symmetric positive definite
            # mtype =-2: real and symmetric indefinite, 
