### 2024-09-24 8eff506
Enhance the docstring of `FieldContainer.extract()`
