### 2024-10-15 #212 @tkoyama010
@cmarmo I have completed the review by taking over [your checklist](https://github.com/pyOpenSci/software-submission/issues/212#issuecomment-2365311571). Please tell us the flow from here.
