### 2024-10-15 #212 @cmarmo
Thank you @tkoyama010 for following-up! Now the hard part will start... We need to [look for reviewers](https://www.pyopensci.org/software-peer-review/how-to/editors-guide.html#identify-scientific-python-package-reviewers) for the package.
I have added you to the editor's Slack channel, so feel free to chat there: we can check in our pool of reviewers if some of them fit this package.
