### 2024-10-15 29c0931
Add a docstring for the multi-point-constraint
