### 2024-10-16 4626afb
Update _multipoint.py
