### 2024-10-16 c21538c
Update _multipoint.py
