### 2024-10-16 6f59beb
Docs: remove howto/rbe2, now located in docstring of `MultiPointConstraint`
