+++ docs/howto.rst
@@ -8,7 +8,6 @@ How-To Guides
    :caption: How-To:
    
    howto/meshgen
-   howto/rbe2
    howto/mixed
    howto/axi
    howto/solvers
