### 2024-10-18 ecc22c1
Add docstring for `MultiPointContact`
