### 2024-10-18 90db4e8
Update _multipoint.py
