### 2024-10-18 19bf43e
Add an optional multiplier in `Laplace(multiplier=1.0)`
