### 2024-10-21 a3e74ad
Add optional kwargs in `math.transpose(**kwargs)`
