+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. The format 
 - Add the hessian of the element shape functions for a quadratic quad element `QuadraticQuad.hessian()`.
 - Add the `order`-argument to `FieldContainer.extract(order="C")` as well as for `Field`, `FieldAxisymmetric`, `FieldPlaneStrain` to return C-contiguous arrays by default.
 - Add an optional multiplier to `Laplace(multiplier=1.0)`.
+- Add optional keyword-arguments to `math.transpose(**kwargs)` to support optional `out` and `order`-keywords.
 
 ### Changed
 - Change default `np.einsum(..., order="K")` to `np.einsum(..., order="C")` in the methods of `Field`, `FieldAxisymmetric`, `FieldPlaneStrain` and `FieldContainer`.
