### 2024-10-21 e04456a
Add optional kwargs in `math.transpose(**kwargs)` (#865)
* Add optional kwargs in `math.transpose(**kwargs)`

* format black
