### 2024-10-21 ff5dd49
Fix the number of points for non-disconnected dual-meshes
