### 2024-10-21 c07d0a9
Fix the number of points for non-disconnected dual-meshes (#867)
