### 2024-10-23 #212 @adtzlr
Hey @cmarmo and @tkoyama010!

In the meantime we prepared a draft paper 📝 for JOSS (meanwhile located in a private repo 🔒). Do you think it is okay to already include it in the FElupe-repo or better wait until the PyOpenSci review is finished?

Why do you require that the paper must be *...in the package root or in `inst/`*? There are a lot of published papers which have their paper located at subfolders, like `paper/paper.md` (I checked a few of the [latest published](https://joss.theoj.org/papers/published) papers).

🙏🏻 Thank you! 🙏🏻
