### 2024-10-24 6704ed0
Enhance the docstring of the linear-elastic material
add a warning on large-strain analyses
