### 2024-10-25 9e07aa8
Fix wrong result of `ConstitutiveMaterial.plot()` if any stretch <= 0 (#872)
this raises an error now to avoid the wrong results
