### 2024-10-26 #212 @cmarmo
> In the meantime we prepared a draft paper 📝 for JOSS (meanwhile located in a private repo 🔒). Do you think it is okay to already include it in the FElupe-repo or better wait until the PyOpenSci review is finished?

Please, add the paper to the repo, so our reviewers can have a look at it, this will speed up things for the JOSS submission.

> Why do you require that the paper must be _...in the package root or in `inst/`_? There are a lot of published papers which have their paper located at subfolders, like `paper/paper.md` (I checked a few of the [latest published](https://joss.theoj.org/papers/published) papers).

euh .... :thinking: I cannot find an explanation for that... I suggest you to follow [JOSS requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements), while we clarify this.

Thank you!
