### 2024-10-28 d1ff525
Add `math.inplane(A, vectors)`
