+++ src/felupe/math/__init__.py
@@ -33,6 +33,7 @@ from ._tensor import (
     eigvalsh,
     equivalent_von_mises,
     identity,
+    inplane,
     inv,
     majortranspose,
     ravel,
@@ -74,6 +75,7 @@ __all__ = [
     "eigvalsh",
     "equivalent_von_mises",
     "identity",
+    "inplane",
     "inv",
     "majortranspose",
     "ravel",
