### 2024-10-28 0124998
Update math.rst
