### 2024-10-28 03d6506
Add `RegionBoundary.tangents` and `math.inplane()` (#873)
* Add `RegionBoundary.tangents` attribute

* Add `math.inplane(A, vectors)`

* Update CHANGELOG.md
