### 2024-10-28 e942040
Enhance the docstring for `RegionBoundary`
