### 2024-10-29 f9fec84
Generalize `math.inplane()` to non-symmetric tensors
