+++ docs/felupe/constitution/jax.rst
@@ -1,0 +1,21 @@
+.. _felupe-api-constitution-jax:
+
+JAX-based Materials
+~~~~~~~~~~~~~~~~~~~
+
+This page contains hyperelastic material model formulations with automatic differentiation using :mod:`jax`. These material model formulations are defined by a strain energy density function.
+
+**Frameworks**
+
+.. currentmodule:: felupe
+
+.. autosummary::
+   
+   constitution.jax.Hyperelastic
+
+**Detailed API Reference**
+
+.. autoclass:: felupe.constitution.jax.Hyperelastic
+   :members:
+   :undoc-members:
+   :inherited-members:
