### 2024-10-31 19022c9
Update test_constitution_jax.py
