+++ tests/test_constitution_jax.py
@@ -69,6 +69,7 @@ def test_hyperelastic_jax():
             I1 = I3 ** (-1 / 3) * jnp.trace(C)
             return C10 * (I1 - 3) + K * (J - 1) ** 2 / 2
 
+        umat = fem.constitution.jax.Hyperelastic(W, C10=0.5, K=2.0, parallel=True)
         umat = fem.constitution.jax.Hyperelastic(W, C10=0.5, K=2.0, jit=True)
         mesh = fem.Cube(n=2)
         region = fem.RegionHexahedron(mesh)
@@ -97,6 +98,8 @@ def test_hyperelastic_jax_statevars():
             I1 = I3 ** (-1 / 3) * jnp.trace(C)
             statevars_new = I1
             return C10 * (I1 - 3) + K * (J - 1) ** 2 / 2, statevars_new
+
+        W.kwargs = {"C10": 0.5}
 
         umat = fem.constitution.jax.Hyperelastic(
             W, C10=0.5, K=2.0, nstatevars=1, jit=True
