### 2024-10-31 df80b5c
Update _hyperelastic.py
