### 2024-10-31 ac460a9
Add `constitution.jax.Hyperelastic` (#876)
* Add `constitution.jax.Hyperelastic`

* Update test_constitution_jax.py

* Update test_constitution_jax.py

* Update test_constitution_jax.py

* Update _hyperelastic.py
