+++ src/felupe/constitution/jax/__init__.py
@@ -1,0 +1,3 @@
+from ._hyperelastic import Hyperelastic, vmap
+
+__all__ = ["Hyperelastic", "vmap"]
