+++ src/felupe/constitution/autodiff/tensortrax/models/__init__.py
@@ -1,0 +1,3 @@
+from . import hyperelastic, lagrange
+
+__all__ = ["hyperelastic", "lagrange"]
