+++ src/felupe/constitution/autodiff/tensortrax/__init__.py
@@ -1,0 +1,13 @@
+from . import models
+from ._hyperelastic import Hyperelastic
+from ._material import Material
+from ._total_lagrange import total_lagrange
+from ._updated_lagrange import updated_lagrange
+
+__all__ = [
+    "Hyperelastic",
+    "models",
+    "total_lagrange",
+    "updated_lagrange",
+    "Material",
+]
