+++ src/felupe/constitution/hyperelasticity/_neo_hooke_nearly_incompressible.py
@@ -18,8 +18,8 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
 
-from ....math import cdya_ik, ddot, det, dot, dya, identity, inv, trace, transpose
-from ..._base import ConstitutiveMaterial
+from ...math import cdya_ik, ddot, det, dot, dya, identity, inv, trace, transpose
+from .._base import ConstitutiveMaterial
 
 
 class NeoHooke(ConstitutiveMaterial):
