+++ src/felupe/constitution/hyperelasticity/_ogden_roxburgh.py
@@ -19,8 +19,8 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 import numpy as np
 from scipy.special import erf
 
-from ....math import dya
-from ..._base import ConstitutiveMaterial
+from ...math import dya
+from .._base import ConstitutiveMaterial
 
 
 class OgdenRoxburgh(ConstitutiveMaterial):
