### 2024-11-01 9d6ead6
Add `felupe.constitution.autodiff.jax.Material`
as an alternative for `MaterialAD` (`felupe.constitution.autodiff.tensortrax.Material`)
