### 2024-11-01 de987bb
test new material class for jax
