### 2024-11-01 4811b26
Add `felupe.constitution.autodiff.jax.Material` (#879)
* Add `felupe.constitution.autodiff.jax.Material`

as an alternative for `MaterialAD` (`felupe.constitution.autodiff.tensortrax.Material`)

* test new material class for jax
