+++ src/felupe/constitution/autodiff/jax/_material.py
@@ -20,22 +20,22 @@ import warnings
 
 import numpy as np
 
-from ..._material import Material
-from ._tools import total_lagrange, vmap2
+from ..._material import Material as MaterialDefault
+from ._tools import vmap2
 
 
-class Hyperelastic(Material):
-    r"""A hyperelastic material definition with a given function for the strain energy
-    density function per unit undeformed volume with Automatic Differentiation provided
-    by :mod:`jax`.
+class Material(MaterialDefault):
+    r"""A material definition with a given function for the partial derivative of the
+    strain energy function w.r.t. the deformation gradient tensor with Automatic
+    Differentiation provided by :mod:`jax`.
 
     Parameters
     ----------
     fun : callable
-        A strain energy density function in terms of the right Cauchy-Green deformation
-        tensor :math:`\boldsymbol{C}`. Function signature must be
-        ``fun = lambda C, **kwargs: psi`` for functions without state variables and
-        ``fun = lambda C, statevars, **kwargs: [psi, statevars_new]`` for functions
+        A gradient of the strain energy density function w.r.t. the deformation gradient
+        tensor :math:`\boldsymbol{F}`. Function signature must be
+        ``fun = lambda F, **kwargs: P`` for functions without state variables and
+        ``fun = lambda F, statevars, **kwargs: [P, statevars_new]`` for functions
         with state variables. It is important to use only differentiable math-functions
         from :mod:`jax`.
     nstatevars : int, optional
@@ -43,66 +43,68 @@ class Hyperelastic(Material):
     jit : bool, optional
         A flag to invoke just-in-time compilation (default is True).
     parallel : bool, optional
-        A flag to invoke threaded strain energy density function evaluations (default
-        is False). Not implemented.
+        A flag to invoke threaded function evaluations (defaultnis False). Not
+        implemented.
     **kwargs : dict, optional
-        Optional keyword-arguments for the strain energy density function.
+        Optional keyword-arguments for the gradient of the strain energy density
+        function.
 
     Notes
     -----
-    The strain energy density function :math:`\psi` must be given in terms of the right
-    Cauchy-Green deformation tensor
-    :math:`\boldsymbol{C} = \boldsymbol{F}^T \boldsymbol{F}`.
+    The gradient of the strain energy density function
+    :math:`\frac{\partial \psi}{\partial \boldsymbol{F}}` must be given in terms of the
+    deformation gradient tensor :math:`\boldsymbol{F}`.
 
     ..  warning::
         It is important to use only differentiable math-functions from :mod:`jax`!
 
-    Take this minimal code-block as template
-
-    ..  math::
-
-        \psi = \psi(\boldsymbol{C})
+    Take this code-block as template
 
     ..  code-block::
 
         import felupe as fem
         import jax.numpy as jnp
 
-        def neo_hooke(C, mu):
-            "Strain energy function of the Neo-Hookean material formulation."
-            return mu / 2 * (jnp.linalg.det(C) ** (-1/3) * jnp.trace(C) - 3)
+        def neo_hooke(F, mu):
+            "First Piola-Kirchhoff stress of the Neo-Hookean material formulation."
 
-        umat = fem.constitution.autodiff.jax.Hyperelastic(neo_hooke, mu=1)
+            C = F.T @ F
[…]
