+++ src/felupe/constitution/autodiff/jax/__init__.py
@@ -1,3 +1,5 @@
-from ._hyperelastic import Hyperelastic, vmap
+from ._hyperelastic import Hyperelastic
+from ._material import Material
+from ._tools import vmap
 
-__all__ = ["Hyperelastic", "vmap"]
+__all__ = ["Hyperelastic", "Material", "vmap"]
