### 2024-11-01 b87862a
Update lagrange.rst
