### 2024-11-02 1422ab4
Add `morph()` for JAX-based `Material`
