### 2024-11-02 7acda9a
Update CHANGELOG.md
