### 2024-11-02 39738ea
Add `morph()` for JAX-based `Material` (#880)
* Add `morph()` for JAX-based `Material`

* remove Python 3.9 support (EOL)

* Update index.rst

* update to Python 3.9+ (docs)

* Update _morph.py

* Update _morph.py

* Update _material.py

* Create .gitkeep

* Update CHANGELOG.md

* Delete .gitkeep

* Create __init__.py
