+++ src/felupe/constitution/autodiff/jax/models/lagrange/_morph.py
@@ -1,0 +1,228 @@
+# -*- coding: utf-8 -*-
+"""
+This file is part of FElupe.
+
+FElupe is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+FElupe is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
+"""
+
+
+def morph(F, statevars, p):
+    r"""Second Piola-Kirchhoff stress tensor of the
+    `MORPH <https://doi.org/10.1016/s0749-6419(02)00091-8>`_ model formulation [1]_.
+
+    Parameters
+    ----------
+    C : tensortrax.Tensor
+        Right Cauchy-Green deformation tensor.
+    statevars : array
+        Vector of stacked state variables (CTS, C, SA).
+    p : list of float
+        A list which contains the 8 material parameters.
+
+    Notes
+    -----
+    The MORPH material model is implemented as a second Piola-Kirchhoff stress-based
+    formulation with automatic differentiation. The Tresca invariant of the distortional
+    part of the right Cauchy-Green deformation tensor is used as internal state
+    variable, see Eq. :eq:`morph-state`.
+
+    ..  warning::
+        While the `MORPH <https://doi.org/10.1016/s0749-6419(02)00091-8>`_-material
+        formulation captures the Mullins effect and quasi-static hysteresis effects of
+        rubber mixtures very nicely, it has been observed to be unstable for medium- to
+        highly-distorted states of deformation.
+
+    ..  math::
+        :label: morph-state
+
+        \boldsymbol{C} &= \boldsymbol{F}^T \boldsymbol{F}
+
+        I_3 &= \det (\boldsymbol{C})
+
+        \hat{\boldsymbol{C}} &= I_3^{-1/3} \boldsymbol{C}
+
+        \hat{\lambda}^2_\alpha &= \text{eigvals}(\hat{\boldsymbol{C}})
+
+        \hat{C}_T &= \max \left( \hat{\lambda}^2_\alpha - \hat{\lambda}^2_\beta \right)
+
+        \hat{C}_T^S &= \max \left( \hat{C}_T, \hat{C}_{T,n}^S \right)
+
+    A sigmoid-function is used inside the deformation-dependent variables
+    :math:`\alpha`, :math:`\beta` and :math:`\gamma`, see Eq. :eq:`morph-sigmoid`.
+
+    ..  math::
+        :label: morph-sigmoid
+
+        f(x) &= \frac{1}{\sqrt{1 + x^2}}
+
+        \alpha &= p_1 + p_2 \ f(p_3\ C_T^S)
+
+        \beta &= p_4\ f(p_3\ C_T^S)
+
+        \gamma &= p_5\ C_T^S\ \left( 1 - f\left(\frac{C_T^S}{p_6}\right) \right)
+
+    The rate of deformation is described by the Lagrangian tensor and its Tresca-
+    invariant, see Eq. :eq:`morph-rate-of-deformation`.
+
+    ..  note::
+        It is important to evaluate the incremental right Cauchy-Green tensor by the
[…]
