### 2024-11-02 c13a0d9
Simplify constitution
