+++ src/felupe/constitution/tensortrax/models/lagrange/microsphere/_framework_affine.py
@@ -1,7 +1,7 @@
 from tensortrax.math import einsum, sqrt, trace
 from tensortrax.math.linalg import det, inv
 
-from .......quadrature import BazantOh
+from ......quadrature import BazantOh
 from ...._total_lagrange import total_lagrange
 
 
