### 2024-11-02 bd1bf63
Simplify constitution (#881)
