### 2024-11-02 26c2043
Enhance the docs on automatic differentiation
