+++ src/felupe/constitution/tensortrax/_hyperelastic.py
@@ -64,13 +64,15 @@ class Hyperelastic(Material):
 
     ..  code-block::
 
+        import felupe as fem
+        import felupe.constitution.tensortrax as mat
         import tensortrax.math as tm
 
         def neo_hooke(C, mu):
             "Strain energy function of the Neo-Hookean material formulation."
             return mu / 2 * (tm.linalg.det(C) ** (-1/3) * tm.trace(C) - 3)
 
-        umat = fem.Hyperelastic(neo_hooke, mu=1)
+        umat = mat.Hyperelastic(neo_hooke, mu=1)
 
     and this code-block for material formulations with state variables.
 
@@ -80,6 +82,8 @@ class Hyperelastic(Material):
 
     ..  code-block::
 
+        import felupe as fem
+        import felupe.constitution.tensortrax as mat
         import tensortrax.math as tm
 
         def viscoelastic(C, Cin, mu, eta, dtime):
@@ -98,9 +102,7 @@ class Hyperelastic(Material):
             # strain energy function and state variable
             return mu / 2 * (I1 - 3), tm.special.triu_1d(Ci)
 
-        umat = fem.Hyperelastic(
-            viscoelastic, mu=1, eta=1, dtime=1, nstatevars=6
-        )
+        umat = mat.Hyperelastic(viscoelastic, mu=1, eta=1, dtime=1, nstatevars=6)
 
     ..  note::
         See the `documentation of tensortrax <https://tensortrax.readthedocs.io>`_
@@ -114,13 +116,14 @@ class Hyperelastic(Material):
         :context:
 
         >>> import felupe as fem
+        >>> import felupe.constitution.jax as mat
         >>> import tensortrax.math as tm
         >>>
         >>> def neo_hooke(C, mu):
         ...     "Strain energy function of the Neo-Hookean material formulation."
         ...     return mu / 2 * (tm.linalg.det(C) ** (-1/3) * tm.trace(C) - 3)
         >>>
-        >>> umat = fem.Hyperelastic(neo_hooke, mu=1)
+        >>> umat = mat.Hyperelastic(neo_hooke, mu=1)
         >>> ax = umat.plot(incompressible=True)
 
     ..  pyvista-plot::
