### 2024-11-02 bf5e7be
Update CHANGELOG.md
