### 2024-11-02 9b57af1
Update ex13_morph-rubber-wheel.py
switch to JAX backend
