+++ examples/ex13_morph-rubber-wheel.py
@@ -6,9 +6,9 @@ This example contains a simulation of a rotating rubber wheel in plane strain wi
 [1]_. While the rotation is increased, a constant vertical compression is applied to the
 rubber wheel by a frictionless contact on the bottom. The vertical reaction force is
 then carried out for the rotation angles. The MORPH material model is implemented as a
-second Piola-Kirchhoff stress-based formulation with automatic differentiation. The
-Tresca invariant of the distortional part of the right Cauchy-Green deformation tensor
-is used as internal state variable, see Eq. :eq:`morph-state`.
+second Piola-Kirchhoff stress-based formulation with automatic differentiation (JAX).
+The Tresca invariant of the distortional part of the right Cauchy-Green deformation
+tensor is used as internal state variable, see Eq. :eq:`morph-state`.
 
 ..  warning::
     While the `MORPH <https://doi.org/10.1016/s0749-6419(02)00091-8>`_-material
@@ -99,86 +99,87 @@ this example.
 """
 # sphinx_gallery_thumbnail_number = -1
 import numpy as np
-import tensortrax.math as tm
-
 import felupe as fem
+import felupe.constitution.jax as mat
+from jax.scipy.linalg import expm
+import jax.numpy as jnp
 
 
-@fem.total_lagrange
-def morph(F, statevars_old, p):
+def morph(F, statevars, p):
     "MORPH material model formulation."
 
     # right Cauchy-Green deformation tensor
     C = F.T @ F
 
     # extract old state variables
-    CTSn = tm.array(statevars_old[0], like=C[0, 0])
-    Cn = tm.special.from_triu_1d(statevars_old[1:7], like=C)
-    SAn = tm.special.from_triu_1d(statevars_old[7:13], like=C)
+    CTSn = statevars[0]
+    from_triu = lambda C: C[jnp.array([[0, 1, 2], [1, 3, 4], [2, 4, 5]])]
+    Cn = from_triu(statevars[1:7])
+    SAn = from_triu(statevars[7:13])
 
     # distortional part of right Cauchy-Green deformation tensor
-    I3 = tm.linalg.det(C)
+    I3 = jnp.linalg.det(C)
     CG = C * I3 ** (-1 / 3)
 
     # inverse of and incremental right Cauchy-Green deformation tensor
-    invC = tm.linalg.inv(C)
+    invC = jnp.linalg.inv(C)
     dC = C - Cn
 
     # eigenvalues of right Cauchy-Green deformation tensor (sorted in ascending order)
-    λCG = tm.linalg.eigvalsh(CG)
+    ε = jnp.diag(jnp.array([1e-4, -1e-4, 0]))
+    eigvalsh_ε = lambda C: jnp.linalg.eigvalsh(C + ε)
+    λCG = eigvalsh_ε(CG)
 
     # Tresca invariant of distortional part of right Cauchy-Green deformation tensor
     CTG = λCG[-1] - λCG[0]
 
     # maximum Tresca invariant in load history
-    CTS = tm.maximum(CTG, CTSn)
+    CTS = jnp.maximum(CTG, CTSn)
 
-    def f(x):
+    def sigmoid(x):
         "Algebraic sigmoid function."
-        return 1 / tm.sqrt(1 + x**2)
+        return 1 / jnp.sqrt(1 + x**2)
 
     # material parameters
-    α = p[0] + p[1] * f(p[2] * CTS)
[…]
