### 2024-11-02 0947f51
Fix JAX intersphinx
switch from stable to latest
