+++ docs/conf.py
@@ -79,7 +79,7 @@ sphinx_gallery_conf = {
     "within_subsection_order": FileNameSortKey,
 }
 intersphinx_mapping = {
-    "jax": ("https://jax.readthedocs.io/en/stable/", None),
+    "jax": ("https://jax.readthedocs.io/en/latest/", None),
     "matplotlib": ("https://matplotlib.org/stable/", None),
     "numpy": ("https://numpy.org/doc/stable/", None),
     "python": ("https://docs.python.org/3/", None),
