### 2024-11-02 6d09742
enhance the examples
