### 2024-11-02 c2dc5d9
force static plots in docstrings
try to make readthedocs run faster
