+++ src/felupe/mesh/_container.py
@@ -51,6 +51,7 @@ class MeshContainer:
     --------
     .. pyvista-plot::
        :include-source: True
+       :force_static:
 
        >>> import felupe as fem
        >>>
