### 2024-11-03 2542936
Add Total- and Updated-Lagrange for JAX
