+++ src/felupe/constitution/jax/__init__.py
@@ -2,5 +2,14 @@ from . import models
 from ._hyperelastic import Hyperelastic
 from ._material import Material
 from ._tools import vmap
+from ._total_lagrange import total_lagrange
+from ._updated_lagrange import updated_lagrange
 
-__all__ = ["Hyperelastic", "Material", "models", "vmap"]
+__all__ = [
+    "Hyperelastic",
+    "Material",
+    "models",
+    "total_lagrange",
+    "updated_lagrange",
+    "vmap",
+]
