### 2024-11-03 33bfb9f
Add Mooney-Rivlin material for JAX
