+++ src/felupe/constitution/jax/models/hyperelastic/_mooney_rivlin.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+"""
+This file is part of FElupe.
+
+FElupe is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+FElupe is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
+"""
+from functools import wraps
+
+from ....tensortrax.models.hyperelastic import mooney_rivlin as mr
+
+
+@wraps(mr)
+def mooney_rivlin(C, C10, C01):
+    from jax.numpy import trace
+    from jax.numpy.linalg import det
+
+    J3 = det(C) ** (-1 / 3)
+    I1 = J3 * trace(C)
+    I2 = (I1**2 - J3**2 * trace(C @ C)) / 2
+    return C10 * (I1 - 3) + C01 * (I2 - 3)
