+++ src/felupe/constitution/jax/models/hyperelastic/__init__.py
@@ -1,0 +1,5 @@
+from ._mooney_rivlin import mooney_rivlin
+
+__all__ = [
+    "mooney_rivlin",
+]
