+++ examples/ex05_rubber-metal-bushing.py
@@ -24,6 +24,7 @@ Elastic bearing with torsional loading
 An elastic bearing is subjected to combined multiaxial radial-torsional-cardanic
 loading. First the meshes for the rubber and the metal sheet rings are created.
 """
+
 # sphinx_gallery_thumbnail_number = -2
 import numpy as np
 import pypardiso
