+++ docs/tutorial/examples/extut03_building_blocks.py
@@ -23,6 +23,7 @@ Start setting up a problem in FElupe by the creation of a numeric :class:`~felup
 .. image:: examples/extut03_building_blocks_sketch.svg
    :width: 600px
 """
+
 # sphinx_gallery_thumbnail_number = -1
 import felupe as fem
 
