### 2024-11-03 7f3c2ab
Update jax.rst
