### 2024-11-03 dcc7939
Update test_constitution_jax.py
