+++ src/felupe/constitution/jax/models/hyperelastic/_yeoh.py
@@ -17,15 +17,13 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 """
 from functools import wraps
 
-from ....tensortrax.models.hyperelastic import mooney_rivlin as mooney_rivlin_docstring
+from ....tensortrax.models.hyperelastic import yeoh as yeoh_docstring
 
 
-@wraps(mooney_rivlin_docstring)
-def mooney_rivlin(C, C10, C01):
+@wraps(yeoh_docstring)
+def yeoh(C, C10, C20, C30):
     from jax.numpy import trace
     from jax.numpy.linalg import det
 
-    J3 = det(C) ** (-1 / 3)
-    I1 = J3 * trace(C)
-    I2 = (I1**2 - J3**2 * trace(C @ C)) / 2
-    return C10 * (I1 - 3) + C01 * (I2 - 3)
+    I1 = det(C) ** (-1 / 3) * trace(C)
+    return C10 * (I1 - 3) + C20 * (I1 - 3) ** 2 + C30 * (I1 - 3) ** 3
