+++ src/felupe/constitution/tensortrax/models/hyperelastic/_third_order_deformation.py
@@ -27,7 +27,7 @@ def third_order_deformation(C, C10, C01, C11, C20, C30):
 
     Parameters
     ----------
-    C : tensortrax.Tensor
+    C : tensortrax.Tensor or jax.Array
         Right Cauchy-Green deformation tensor.
     C10 : float
         Material parameter associated to the linear term of the first invariant.
@@ -78,14 +78,26 @@ def third_order_deformation(C, C10, C01, C11, C20, C30):
 
     Examples
     --------
+    First, choose the desired automatic differentiation backend
 
     ..  pyvista-plot::
         :context:
 
-        >>> import felupe as fem
-        >>>
-        >>> umat = fem.Hyperelastic(
-        ...     fem.third_order_deformation, C10=0.5, C01=0.1, C11=0.01, C20=-0.1, C30=0.02
+        >>> # import felupe.constitution.jax as mat
+        >>> import felupe.constitution.tensortrax as mat
+
+    and create the hyperelastic material.
+
+    ..  pyvista-plot::
+        :context:
+
+        >>> umat = mat.Hyperelastic(
+        ...     mat.models.hyperelastic.third_order_deformation,
+        ...     C10=0.5,
+        ...     C01=0.1,
+        ...     C11=0.01,
+        ...     C20=-0.1,
+        ...     C30=0.02,
         ... )
         >>> ax = umat.plot(incompressible=True)
 
