+++ src/felupe/constitution/jax/models/hyperelastic/__init__.py
@@ -1,5 +1,9 @@
 from ._mooney_rivlin import mooney_rivlin
+from ._third_order_deformation import third_order_deformation
+from ._yeoh import yeoh
 
 __all__ = [
     "mooney_rivlin",
+    "third_order_deformation",
+    "yeoh",
 ]
