### 2024-11-03 af7317f
Update test_constitution_jax.py
