### 2024-11-03 faa638d
Update jax.rst
