+++ tests/test_constitution_jax.py
@@ -127,7 +127,7 @@ def test_material_jax():
         umat = mat.Material(dWdF, C10=0.5, K=2.0, jit=True)
 
         for fun in [dWdF, mat.updated_lagrange(dWdF), mat.total_lagrange(dWdF)]:
-            umat = mat.Material(dWdF, C10=0.5, K=2.0)
+            umat = mat.Material(fun, C10=0.5, K=2.0)
             mesh = fem.Cube(n=2)
             region = fem.RegionHexahedron(mesh)
             field = fem.FieldContainer([fem.Field(region, dim=3)])
