### 2024-11-03 4fdcc12
Update CHANGELOG.md
