### 2024-11-03 eed4ece
Add Total-Lagrange and Updated-Lagrange decorators for JAX (#883)
* Add Total- and Updated-Lagrange for JAX

* Add Mooney-Rivlin material for JAX

* JAX-based MORPH material as Total-Lagrange

* Update _morph_representative_directions.py

enhance docstring example

* format black

* Update jax.rst

* Update test_constitution_jax.py

* Add TOD and YEOH for JAX

* Update test_constitution_jax.py

* Update test_constitution_jax.py

* Update jax.rst

* Update test_constitution_jax.py
