+++ examples/ex02_plate-with-hole.py
@@ -21,6 +21,7 @@ over the hole.
 .. image:: ../../examples/ex02_plate-with-hole_sketch.svg
    :width: 400px
 """
+
 # sphinx_gallery_thumbnail_number = -2
 h = 1
 L = 2
