+++ src/felupe/constitution/jax/models/hyperelastic/_third_order_deformation.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+"""
+This file is part of FElupe.
+
+FElupe is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+FElupe is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
+"""
+from functools import wraps
+
+from ....tensortrax.models.hyperelastic import third_order_deformation as tod_docstring
+
+
+@wraps(tod_docstring)
+def third_order_deformation(C, C10, C01, C11, C20, C30):
+    from jax.numpy import trace
+    from jax.numpy.linalg import det
+
+    J3 = det(C) ** (-1 / 3)
+    I1 = J3 * trace(C)
+    I2 = (I1**2 - J3**2 * trace(C @ C)) / 2
+    return (
+        C10 * (I1 - 3)
+        + C01 * (I2 - 3)
+        + C11 * (I1 - 3) * (I2 - 3)
+        + C20 * (I1 - 3) ** 2
+        + C30 * (I1 - 3) ** 3
+    )
