### 2024-11-04 1c910d6
Add more JAX-based models
