+++ src/felupe/constitution/jax/models/lagrange/microsphere/__init__.py
@@ -1,0 +1,5 @@
+from ._framework_affine import affine_force_statevars
+
+__all__ = [
+    "affine_force_statevars",
+]
