### 2024-11-05 7849c07
Add tests for JAX material
