### 2024-11-05 5c71376
Add more JAX-based models (#884)
* Add more JAX-based models

* Add tests for JAX material

* Update _morph_uniaxial.py
