### 2024-11-05 3bed5b4
Add `isochoric_volumetric_split()` for JAX-hyperelastic material
