### 2024-11-05 eb32edd
Add `isochoric_volumetric_split()` for JAX-hyperelastic material (#885)
