+++ tests/test_constitution_newton.py
@@ -25,21 +25,24 @@ along with Felupe.  If not, see <http://www.gnu.org/licenses/>.
 
 """
 import numpy as np
-import tensortrax as tr
-from tensortrax.math import trace
-from tensortrax.math.linalg import det, inv
-from tensortrax.math.special import dev, from_triu_1d, triu_1d
 
 import felupe as fem
 
 
 def test_visco_newton():
+    import tensortrax as tr
+    from tensortrax.math import trace
+    from tensortrax.math.linalg import det, inv
+    from tensortrax.math.special import from_triu_1d, triu_1d
+
+    import felupe.constitution.tensortrax as mat
+
     mesh = fem.Rectangle(n=3)
     region = fem.RegionQuad(mesh)
     field = fem.FieldContainer([fem.FieldPlaneStrain(region, dim=2)])
     boundaries, loadcase = fem.dof.uniaxial(field, clamped=True)
 
-    @fem.constitution.isochoric_volumetric_split
+    @mat.models.hyperelastic.isochoric_volumetric_split
     def finite_strain_viscoelastic_newton(C, Cin, mu, eta, dtime):
         "Finite Strain Viscoelastic material formulation."
 
@@ -65,11 +68,10 @@ def test_visco_newton():
         # strain energy function and state variable
         return mu / 2 * (I1 - 3), triu_1d(Ci)
 
-    umat = fem.Hyperelastic(
+    umat = mat.Hyperelastic(
         finite_strain_viscoelastic_newton, mu=1.0, eta=1.0, dtime=0.1, nstatevars=6
     )
     solid = fem.SolidBodyNearlyIncompressible(umat, field, bulk=5000)
-    # solid.results.statevars[[0, 3, 5]] += 1
 
     move = fem.math.linsteps([0, 0.3], num=3)
     step = fem.Step(
@@ -79,5 +81,66 @@ def test_visco_newton():
     assert np.all([norm[-1] < 1e-7 for norm in job.fnorms])
 
 
+def test_visco_newton_jax():
+    import jax
+    from jax.numpy import trace
+    from jax.numpy.linalg import det, inv
+    from jax.scipy.optimize import minimize
+
+    import felupe.constitution.jax as mat
+
+    mesh = fem.Rectangle(n=3)
+    region = fem.RegionQuad(mesh)
+    field = fem.FieldContainer([fem.FieldPlaneStrain(region, dim=2)])
+    boundaries, loadcase = fem.dof.uniaxial(field, clamped=True)
+
+    @mat.models.hyperelastic.isochoric_volumetric_split
+    def finite_strain_viscoelastic_newton(C, statevars, mu, eta, dtime):
+        "Finite Strain Viscoelastic material formulation."
+
+        def evolution(Ci, Cin, C, mu, eta, dtime):
+            "Viscoelastic evolution equation."
+            Ci = Ci.reshape(3, 3)
+            residuals = mu / eta * C - (Ci - Cin) / dtime
+            return residuals.ravel() @ residuals.ravel()
+
+        # update of state variables by evolution equation
+        Cin = statevars[:9].reshape(3, 3)
+        Ci = minimize(
+            evolution,
+            x0=Cin.ravel(),
+            args=(Cin, jax.lax.stop_gradient(C), mu, eta, dtime),
+            method="BFGS",
+        ).x.reshape(3, 3)
+        Ci *= det(Ci) ** (-1 / 3)
+
+        # first invariant of elastic part of right Cauchy-Green deformation tensor
+        I1 = trace(C @ inv(Ci))
+
+        # strain energy function and state variable
+        statevars_new = Ci.ravel()
[…]
