+++ src/felupe/constitution/jax/models/hyperelastic/__init__.py
@@ -1,8 +1,10 @@
+from ._helpers import isochoric_volumetric_split
 from ._mooney_rivlin import mooney_rivlin
 from ._third_order_deformation import third_order_deformation
 from ._yeoh import yeoh
 
 __all__ = [
+    "isochoric_volumetric_split",
     "mooney_rivlin",
     "third_order_deformation",
     "yeoh",
