+++ src/felupe/constitution/jax/models/hyperelastic/_helpers.py
@@ -1,30 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-This file is part of FElupe.
-
-FElupe is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-FElupe is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
-"""
-from functools import wraps
-
-
-def isochoric_volumetric_split(fun):
-    """Apply the material formulation only on the isochoric part of the
-    multiplicative split of the deformation gradient."""
-    from jax.numpy.linalg import det
-
-    @wraps(fun)
-    def apply_iso(C, *args, **kwargs):
-        return fun(det(C) ** (-1 / 3) * C, *args, **kwargs)
-
-    return apply_iso
