### 2024-11-05 b213099
Move `constitution.jax.isochoric_volumetric_split()` (#887)
from `constitution.jax.models.hyperelastic.isochoric_volumetric_split()`; also for tensortrax
