+++ src/felupe/constitution/jax/__init__.py
@@ -1,12 +1,13 @@
 from . import models
+from ._helpers import isochoric_volumetric_split, vmap
 from ._hyperelastic import Hyperelastic
 from ._material import Material
-from ._tools import vmap
 from ._total_lagrange import total_lagrange
 from ._updated_lagrange import updated_lagrange
 
 __all__ = [
     "Hyperelastic",
+    "isochoric_volumetric_split",
     "Material",
     "models",
     "total_lagrange",
