### 2024-11-05 da62927
Update CHANGELOG.md
