### 2024-11-06 47ee8f4
Add optional in-place cast `Region.astype(dtype, copy=True)`
