### 2024-11-06 706a638
Add optional in-place cast `Region.astype(dtype, copy=True)` (#888)
* Add optional in-place cast `Region.astype(dtype, copy=True)`

* Update test_dtype.py
