### 2024-11-07 742fb1e
Make the examples run faster on CI/CD
