### 2024-11-07 51237be
Make the examples run faster on CI/CD (#890)
