### 2024-11-07 59ab595
Update ex11_notch-stress.py
use only mesh-points with connected cells
