### 2024-11-08 b08e200
Update ex09_numeric-continuation.py
use less substeps
