### 2024-11-08 5a35862
Update ex09_numeric-continuation.py
