### 2024-11-08 6d31e79
Ex.04: Add y-symmetry
and use less continuation steps
