### 2024-11-08 61c5b63
Add the `miehe_goektepe_lulei` model for JAX
