### 2024-11-08 e232e53
Add the `miehe_goektepe_lulei` model for JAX (#892)
