+++ src/felupe/constitution/jax/models/hyperelastic/microsphere/__init__.py
@@ -1,0 +1,19 @@
+from ._chain import langevin, linear
+from ._framework_affine import (
+    affine_stretch,
+    affine_stretch_statevars,
+    affine_tube,
+    affine_tube_statevars,
+)
+from ._framework_nonaffine import nonaffine_stretch, nonaffine_tube
+
+__all__ = [
+    "affine_stretch",
+    "affine_stretch_statevars",
+    "affine_tube",
+    "affine_tube_statevars",
+    "linear",
+    "langevin",
+    "nonaffine_stretch",
+    "nonaffine_tube",
+]
