### 2024-11-08 29571c1
Ex. 08: Switch from matADI to JAX
