### 2024-11-09 7d0ed64
Update ex08_shear.py
