### 2024-11-09 15c9668
Update ex08_shear.py
