### 2024-11-09 2721c1b
Make `tensortrax` optional (#894)
* Make `tensortrax` optional

* remove docstring of test-files

* Update test_constitution_jax.py
