+++ src/felupe/__init__.py
@@ -19,7 +19,6 @@ from .constitution import (
     AreaChange,
     CompositeMaterial,
     ConstitutiveMaterial,
-    Hyperelastic,
     Laplace,
     LinearElastic,
     LinearElasticLargeStrain,
@@ -28,7 +27,6 @@ from .constitution import (
     LinearElasticPlasticIsotropicHardening,
     LineChange,
     Material,
-    MaterialAD,
     MaterialStrain,
     NearlyIncompressible,
     NeoHooke,
@@ -39,29 +37,8 @@ from .constitution import (
     ViewMaterialIncompressible,
     VolumeChange,
     Volumetric,
-    alexander,
-    anssari_benam_bucchi,
-    arruda_boyce,
-    constitutive_material,
-    extended_tube,
-    finite_strain_viscoelastic,
-    isochoric_volumetric_split,
     linear_elastic,
     linear_elastic_plastic_isotropic_hardening,
-    lopez_pamies,
-    miehe_goektepe_lulei,
-    mooney_rivlin,
-    morph,
-    morph_representative_directions,
-    neo_hooke,
-    ogden,
-    ogden_roxburgh,
-    saint_venant_kirchhoff,
-    third_order_deformation,
-    total_lagrange,
-    updated_lagrange,
-    van_der_waals,
-    yeoh,
 )
 from .dof import Boundary
 from .element import ArbitraryOrderLagrange as ArbitraryOrderLagrangeElement
@@ -181,10 +158,6 @@ __all__ = [
     "NearlyIncompressible",
     "Material",
     "MaterialStrain",
-    "Hyperelastic",
-    "total_lagrange",
-    "updated_lagrange",
-    "MaterialAD",
     "ViewMaterial",
     "ViewMaterialIncompressible",
     "ConstitutiveMaterial",
@@ -192,24 +165,6 @@ __all__ = [
     "VolumeChange",
     "linear_elastic",
     "linear_elastic_plastic_isotropic_hardening",
-    "alexander",
-    "anssari_benam_bucchi",
-    "arruda_boyce",
-    "extended_tube",
-    "finite_strain_viscoelastic",
-    "isochoric_volumetric_split",
-    "lopez_pamies",
-    "miehe_goektepe_lulei",
-    "mooney_rivlin",
-    "morph",
-    "morph_representative_directions",
-    "neo_hooke",
-    "ogden",
-    "ogden_roxburgh",
-    "saint_venant_kirchhoff",
-    "third_order_deformation",
-    "van_der_waals",
-    "yeoh",
     "Boundary",
     "ArbitraryOrderLagrangeElement",
     "BiQuadraticQuad",
@@ -286,3 +241,58 @@ __all__ = [
     "ViewSolid",
     "runs_on",
 ]
+
+try:
+    from .constitution import (
+        Hyperelastic,
+        MaterialAD,
+        alexander,
+        anssari_benam_bucchi,
+        arruda_boyce,
+        constitutive_material,
+        extended_tube,
+        finite_strain_viscoelastic,
+        isochoric_volumetric_split,
+        lopez_pamies,
+        miehe_goektepe_lulei,
+        mooney_rivlin,
+        morph,
+        morph_representative_directions,
+        neo_hooke,
+        ogden,
+        ogden_roxburgh,
+        saint_venant_kirchhoff,
+        third_order_deformation,
+        total_lagrange,
+        updated_lagrange,
+        van_der_waals,
+        yeoh,
+    )
+
+    __all_tensortrax = [
+        "Hyperelastic",
+        "total_lagrange",
+        "updated_lagrange",
+        "MaterialAD",
+        "alexander",
+        "anssari_benam_bucchi",
+        "arruda_boyce",
[…]
