+++ src/felupe/constitution/jax/__init__.py
@@ -1,16 +1,20 @@
-from . import models
-from ._helpers import isochoric_volumetric_split, vmap
-from ._hyperelastic import Hyperelastic
-from ._material import Material
-from ._total_lagrange import total_lagrange
-from ._updated_lagrange import updated_lagrange
+try:
+    from . import models
+    from ._helpers import isochoric_volumetric_split, vmap
+    from ._hyperelastic import Hyperelastic
+    from ._material import Material
+    from ._total_lagrange import total_lagrange
+    from ._updated_lagrange import updated_lagrange
 
-__all__ = [
-    "Hyperelastic",
-    "isochoric_volumetric_split",
-    "Material",
-    "models",
-    "total_lagrange",
-    "updated_lagrange",
-    "vmap",
-]
+    __all__ = [
+        "Hyperelastic",
+        "isochoric_volumetric_split",
+        "Material",
+        "models",
+        "total_lagrange",
+        "updated_lagrange",
+        "vmap",
+    ]
+
+except ModuleNotFoundError:
+    __all__ = []
