+++ src/felupe/constitution/jax/models/hyperelastic/microsphere/_chain.py
@@ -1,9 +1,11 @@
+from jax.numpy import log, sinh, sqrt
+
+
 def langevin(stretch, mu, N):
     """Langevin model given by the free energy of a single chain as a function of the
     stretch (assuming a complex valued logarithm). The inverse Langevin function is
     defined by a Padé approximation.
     """
-    from jax.numpy import log, sinh, sqrt
 
     x = stretch / sqrt(N)
     L = x * (3 - x**2) / (1 - x**2)

[…]
