### 2024-11-09 f4ac6f9
Test the import in an env with no autodiff
