+++ tox.ini
@@ -15,3 +15,5 @@ commands =
 
 [testenv:test]
 extras = test
+commands =
+    pytest tests/test_import.py {posargs}
