+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py3
+envlist = py3, core
 isolated_build = True
 
 [tool:pytest]
@@ -13,7 +13,7 @@ extras = all
 commands =
     pytest tests {posargs}
 
-[testenv:test]
+[testenv:core]
 extras = test
 commands =
     pytest tests/test_import.py {posargs}
