+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py3, core
+envlist = core, py3
 isolated_build = True
 
 [tool:pytest]
@@ -14,6 +14,5 @@ commands =
     pytest tests {posargs}
 
 [testenv:core]
-extras = test
 commands =
     pytest tests/test_import.py {posargs}
