+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py3
+envlist = core, py3
 isolated_build = True
 
 [tool:pytest]
@@ -13,5 +13,6 @@ extras = all
 commands =
     pytest tests {posargs}
 
-[testenv:test]
-extras = test
+[testenv:core]
+commands =
+    pytest tests/test_import.py {posargs}
