+++ pyproject.toml
@@ -79,8 +79,6 @@ parallel = ["einsumt"]
 progress = ["tqdm"]
 plot = ["matplotlib"]
 view = ["pyvista[jupyter]"]
-
-test = ["felupe[io,autodiff,plot]"]
 all = ["felupe[io,autodiff,parallel,plot,progress,view]"]
 
 [tool.setuptools.dynamic]
