### 2024-11-09 3358876
Remove matadi from examples in docstrings
