+++ src/felupe/constitution/_base.py
@@ -460,12 +460,12 @@ def constitutive_material(Material, name=None):
         ...
         ...     def gradient(self, x):
         ...         F, statevars = x[0], x[-1]
-        ...         dWdF = self.kwargs["a"] * fem.math.identity(F)
+        ...         dWdF = self.kwargs["a"] * np.eye(3).reshape(3, 3, 1, 1)
         ...         return [dWdF, statevars]
         ...
         ...     def hessian(self, x, **kwargs):
         ...         F, statevars = x[0], x[-1]
-        ...         d2WdFdF = self.kwargs["a"] * np.zeros((3, 3, 3, 3, *F.shape[2:]))
+        ...         d2WdFdF = self.kwargs["a"] * np.zeros((3, 3, 3, 3, 1, 1))
         ...         return [d2WdFdF]
         >>>
         >>> MyMaterial = fem.constitutive_material(MyMaterialFormulation)
