### 2024-11-10 0d5bdfb
Attempt to enable `jax.Hyperelastic(..., parallel=True)` with `jax.pmap`
