### 2024-11-10 9e5fdc9
Enhance the docs on `parallel=True` with JAX
