### 2024-11-10 48dc44a
Update test_constitution_jax.py
