+++ tests/test_constitution_jax.py
@@ -83,9 +83,7 @@ def test_hyperelastic_jax_statevars():
         I3 = jnp.linalg.det(C)
         J = jnp.sqrt(I3)
         I1 = I3 ** (-1 / 3) * jnp.trace(C)
-        statevars_new = I1.reshape(
-            1,
-        )
+        statevars_new = statevars.at[0].set(I1)
         return C10 * (I1 - 3) + K * (J - 1) ** 2 / 2, statevars_new
 
     W.kwargs = {"C10": 0.5}
@@ -134,9 +132,7 @@ def test_material_jax_statevars():
         dev = lambda C: C - jnp.trace(C) / 3 * jnp.eye(3)
 
         P = 2 * C10 * F @ dev(Cu) @ jnp.linalg.inv(C)
-        statevars_new = J.reshape(
-            1,
-        )
+        statevars_new = statevars.at[0].set(J)
         return P + K * (J - 1) * J * jnp.linalg.inv(C), statevars_new
 
     dWdF.kwargs = {"C10": 0.5}
