### 2024-11-10 c8bb5df
Enable `jax.Hyperelastic(..., parallel=True)` with `jax.pmap` on quadrature-points axis (#896)
* Attempt to enable `jax.Hyperelastic(..., parallel=True)` with `jax.pmap`

* Enhance the docs on `parallel=True` with JAX

* Update test_constitution_jax.py
