+++ src/felupe/constitution/jax/_hyperelastic.py
@@ -44,8 +44,9 @@ class Hyperelastic(Material):
     jit : bool, optional
         A flag to invoke just-in-time compilation (default is True).
     parallel : bool, optional
-        A flag to invoke threaded strain energy density function evaluations (default
-        is False). Not implemented.
+        A flag to invoke parallel strain energy density function evaluations (default
+        is False). If True, the quadrature points are executed in parallel. The number
+        of devices must be greater or equal the number of quadrature points per cell.
     **kwargs : dict, optional
         Optional keyword-arguments for the strain energy density function.
 
@@ -86,7 +87,7 @@ class Hyperelastic(Material):
 
         import felupe as fem
         import felupe.constitution.jax as mat
-        import jax.numpy as np
+        import jax.numpy as jnp
 
         def viscoelastic(C, Cin, mu, eta, dtime):
             "Finite strain viscoelastic material formulation."
@@ -149,9 +150,6 @@ class Hyperelastic(Material):
         has_aux = nstatevars > 0
         self.fun = as_total_lagrange(fun)
 
-        if parallel:
-            warnings.warn("Parallel execution is not implemented.")
-
         keyword_args = kwargs
         if hasattr(fun, "kwargs"):
             keyword_args = {**fun.kwargs, **keyword_args}
@@ -163,14 +161,30 @@ class Hyperelastic(Material):
             **keyword_args,
         )
 
-        kwargs_jax = dict(in_axes=-1, out_axes=-1)
+        in_axes = out_axes_grad = [2, 3]
         if nstatevars > 0:
-            kwargs_jax["in_axes"] = (-1, -1)
+            in_axes = out_axes_grad = [(2, 1), (3, 2)]
 
-        self._grad = vmap2(jax.grad(self.fun, has_aux=has_aux), **kwargs_jax)
+        out_axes_hess = [4, 5]
+        if nstatevars > 0:
+            out_axes_hess = [(4, 1), (5, 2)]
+
+        methods = [jax.vmap, jax.vmap]
+        if parallel:
+            methods[0] = jax.pmap  # apply on quadrature-points
+            jit = False  # pmap uses jit
+
+        self._grad = vmap2(
+            jax.grad(self.fun, has_aux=has_aux),
+            in_axes=in_axes,
+            out_axes=out_axes_grad,
+            methods=methods,
+        )
         self._hess = vmap2(
             jax.jacfwd(jax.grad(self.fun, has_aux=has_aux), has_aux=has_aux),
-            **kwargs_jax,
+            in_axes=in_axes,
+            out_axes=out_axes_hess,
+            methods=methods,
         )
 
         if jit:
