+++ .github/workflows/upload-codecov.yml
@@ -16,16 +16,16 @@ jobs:
       matrix:
         python-version: ["3.9", "3.12"]
     steps:
-    - uses: actions/setup-python@v4
+    - uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Test with tox
       run: |
         pip install tox
         tox -- --cov felupe --cov-report xml --cov-report term
     - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v4
       if: ${{ matrix.python-version == '3.12' }}
-      env:
-        CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+      uses: codecov/codecov-action@v4
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
