### 2024-11-13 2312941
Add `math.right_cauchy_green_deformation()`
