+++ src/felupe/math/__init__.py
@@ -8,6 +8,7 @@ from ._field import (
     hess,
     interpolate,
     norm,
+    right_cauchy_green_deformation,
     strain,
     strain_stretch_1d,
     values,
@@ -80,6 +81,7 @@ __all__ = [
     "majortranspose",
     "ravel",
     "reshape",
+    "right_cauchy_green_deformation",
     "sym",
     "tovoigt",
     "trace",
