+++ src/felupe/constitution/jax/models/hyperelastic/_third_order_deformation.py
@@ -25,7 +25,6 @@ from ....tensortrax.models.hyperelastic import third_order_deformation as tod_do
 
 @wraps(tod_docstring)
 def third_order_deformation(C, C10, C01, C11, C20, C30):
-
     J3 = det(C) ** (-1 / 3)
     I1 = J3 * trace(C)
     I2 = (I1**2 - J3**2 * trace(C @ C)) / 2
