+++ src/felupe/constitution/jax/models/hyperelastic/_mooney_rivlin.py
@@ -25,7 +25,6 @@ from ....tensortrax.models.hyperelastic import mooney_rivlin as mooney_rivlin_do
 
 @wraps(mooney_rivlin_docstring)
 def mooney_rivlin(C, C10, C01):
-
     J3 = det(C) ** (-1 / 3)
     I1 = J3 * trace(C)
     I2 = (I1**2 - J3**2 * trace(C @ C)) / 2
