### 2024-11-13 e6135d0
Add math-function for the right Cauchy-Green deformation tensor (#898)
* Add `math.right_cauchy_green_deformation()`

* format black

* Update CHANGELOG.md
