### 2024-11-13 1b5716f
remove PyVista-example downloads
to avoid errors on read-the-docs
