### 2024-11-13 0c56e51
Ex. 12: Use a finer mesh
