### 2024-11-13 c5dea62
Update ex18_nonlinear-viscoelasticity-newton.py
