### 2024-11-14 ac39b20
Add the hyperelastic foam model (tensortrax)
