### 2024-11-15 2efc12e
Update _foam.py
