### 2024-11-15 19cffb0
Update _foam.py
