### 2024-11-15 18226de
Check success of transverse-constraint in `ViewMaterial.plot()`
