+++ src/felupe/constitution/tensortrax/models/hyperelastic/__init__.py
@@ -14,7 +14,6 @@ from ._anssari_benam_bucchi import anssari_benam_bucchi
 from ._arruda_boyce import arruda_boyce
 from ._extended_tube import extended_tube
 from ._finite_strain_viscoelastic import finite_strain_viscoelastic
-from ._foam import foam
 from ._lopez_pamies import lopez_pamies
 from ._miehe_goektepe_lulei import miehe_goektepe_lulei
 from ._mooney_rivlin import mooney_rivlin
@@ -23,6 +22,7 @@ from ._neo_hooke import neo_hooke
 from ._ogden import ogden
 from ._ogden_roxburgh import ogden_roxburgh
 from ._saint_venant_kirchhoff import saint_venant_kirchhoff
+from ._storakers import storakers
 from ._third_order_deformation import third_order_deformation
 from ._van_der_waals import van_der_waals
 from ._yeoh import yeoh
@@ -33,7 +33,6 @@ __all__ = [
     "arruda_boyce",
     "extended_tube",
     "finite_strain_viscoelastic",
-    "foam",
     "lopez_pamies",
     "miehe_goektepe_lulei",
     "mooney_rivlin",
@@ -42,6 +41,7 @@ __all__ = [
     "ogden",
     "ogden_roxburgh",
     "saint_venant_kirchhoff",
+    "storakers",
     "third_order_deformation",
     "van_der_waals",
     "yeoh",
@@ -52,7 +52,6 @@ alexander.kwargs = dict(C1=0, C2=0, C3=0, gamma=100, k=0)
 anssari_benam_bucchi.kwargs = dict(mu=0, N=100)
 arruda_boyce.kwargs = dict(C1=0, limit=100)
 extended_tube.kwargs = dict(Gc=0, Ge=0, beta=1, delta=0)
-foam.kwargs = dict(mu=[0, 0], alpha=[2, -2], beta=[1, 1])
 lopez_pamies.kwargs = dict(mu=[0, 0], alpha=[1, 4])
 miehe_goektepe_lulei.kwargs = dict(mu=0, N=100, U=0, p=2, q=2)
 mooney_rivlin.kwargs = dict(C10=0, C01=0)
@@ -61,6 +60,7 @@ neo_hooke.kwargs = dict(mu=0)
 ogden.kwargs = dict(mu=[0, 0], alpha=[2, -2])
 ogden_roxburgh.kwargs = dict(r=100, m=1, beta=0, material=neo_hooke, mu=0)
 saint_venant_kirchhoff.kwargs = dict(mu=0.0, lmbda=0.0)
+storakers.kwargs = dict(mu=[0], alpha=[2], beta=[1])
 third_order_deformation.kwargs = dict(C10=0, C01=0, C11=0, C20=0, C30=0)
 van_der_waals.kwargs = dict(mu=0, beta=0, a=0, limit=100)
 yeoh.kwargs = dict(C10=0, C20=0, C30=0)
