+++ src/felupe/__init__.py
@@ -262,6 +262,7 @@ try:
         ogden,
         ogden_roxburgh,
         saint_venant_kirchhoff,
+        storakers,
         third_order_deformation,
         total_lagrange,
         updated_lagrange,
@@ -289,6 +290,7 @@ try:
         "ogden",
         "ogden_roxburgh",
         "saint_venant_kirchhoff",
+        "storakers",
         "third_order_deformation",
         "van_der_waals",
         "yeoh",
