### 2024-11-16 fbbb00f
Add the hyperelastic foam model also for JAX
