### 2024-11-16 2bf245d
add tests for hyperelastic foam model
