+++ tests/test_constitution_jax.py
@@ -67,6 +67,7 @@ def test_hyperelastic_jax():
         md.yeoh,
         md.third_order_deformation,
         md.miehe_goektepe_lulei,
+        md.storakers,
     ]:
         umat = mat.Hyperelastic(W, **W.kwargs)
         solid = fem.SolidBody(umat=umat, field=field)
