+++ src/felupe/constitution/jax/models/hyperelastic/_storakers.py
@@ -16,7 +16,9 @@ You should have received a copy of the GNU General Public License
 along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 """
 from functools import wraps
-from jax.numpy import array, sum as asum, sqrt
+
+from jax.numpy import array, sqrt
+from jax.numpy import sum as asum
 from jax.numpy.linalg import eigvalsh
 
 from ....tensortrax.models.hyperelastic import storakers as storakers_docstring
@@ -26,9 +28,9 @@ from ....tensortrax.models.hyperelastic import storakers as storakers_docstring
 def storakers(C, mu, alpha, beta):
     λ1, λ2, λ3 = sqrt(eigvalsh(C))
     J = λ1 * λ2 * λ3
-    
+
     μ = array(mu)
     α = array(alpha)
     β = array(beta)
-    
+
     return asum(2 * μ / α**2 * (λ1**α + λ2**α + λ3**α - 3 + (J ** (-α * β) - 1) / β))
