### 2024-11-16 bd0be09
Test the foam model with other parameters
which will fail with incompressibility as initial guess in plotting
