### 2024-11-16 ccabd1f
Test value error in `umat.plot()`
