### 2024-11-16 be970f3
Update test_constitution.py
