### 2024-11-16 35067a9
Add the Storåkers hyperelastic foam model (#900)
* Add the hyperelastic foam model (tensortrax)

* Update _foam.py

* Update _foam.py

* Check success of transverse-constraint in `ViewMaterial.plot()`

* rename the Foam material to Storakers

* Update _storakers.py

* Update __init__.py

* Update __init__.py

* Update _storakers.py

* Add the hyperelastic foam model also for JAX

* add tests for hyperelastic foam model

* Update _storakers.py

* Update jax.rst

* Update tensortrax.rst

* Test the foam model with other parameters

which will fail with incompressibility as initial guess in plott
[…]
