+++ src/felupe/constitution/__init__.py
@@ -67,6 +67,7 @@ try:
         ogden,
         ogden_roxburgh,
         saint_venant_kirchhoff,
+        storakers,
         third_order_deformation,
         van_der_waals,
         yeoh,
@@ -89,6 +90,7 @@ try:
         "ogden",
         "ogden_roxburgh",
         "saint_venant_kirchhoff",
+        "storakers",
         "third_order_deformation",
         "van_der_waals",
         "yeoh",
