### 2024-11-16 6b2d58b
Aktualisieren von _storakers.py
fix docstring typos
