### 2024-11-17 87134c0
Add the Blatz-Ko hyperelastic foam model
