+++ src/felupe/__init__.py
@@ -249,6 +249,7 @@ try:
         alexander,
         anssari_benam_bucchi,
         arruda_boyce,
+        blatz_ko,
         constitutive_material,
         extended_tube,
         finite_strain_viscoelastic,
@@ -278,6 +279,7 @@ try:
         "alexander",
         "anssari_benam_bucchi",
         "arruda_boyce",
+        "blatz_ko",
         "extended_tube",
         "finite_strain_viscoelastic",
         "isochoric_volumetric_split",
