+++ src/felupe/constitution/__init__.py
@@ -58,6 +58,7 @@ try:
         alexander,
         anssari_benam_bucchi,
         arruda_boyce,
+        blatz_ko,
         extended_tube,
         finite_strain_viscoelastic,
         lopez_pamies,
@@ -78,6 +79,7 @@ try:
         "alexander",
         "anssari_benam_bucchi",
         "arruda_boyce",
+        "blatz_ko",
         "extended_tube",
         "finite_strain_viscoelastic",
         "isochoric_volumetric_split",
