+++ src/felupe/constitution/jax/models/hyperelastic/__init__.py
@@ -1,3 +1,4 @@
+from ._blatz_ko import blatz_ko
 from ._miehe_goektepe_lulei import miehe_goektepe_lulei
 from ._mooney_rivlin import mooney_rivlin
 from ._storakers import storakers
@@ -5,6 +6,7 @@ from ._third_order_deformation import third_order_deformation
 from ._yeoh import yeoh
 
 __all__ = [
+    "blatz_ko",
     "miehe_goektepe_lulei",
     "mooney_rivlin",
     "storakers",
@@ -13,6 +15,7 @@ __all__ = [
 ]
 
 # default (stable) material parameters
+blatz_ko.kwargs = dict(mu=1)
 miehe_goektepe_lulei.kwargs = dict(mu=0, N=100, U=0, p=2, q=2)
 mooney_rivlin.kwargs = dict(C10=0, C01=0)
 storakers.kwargs = dict(mu=[0], alpha=[2], beta=[1])
