### 2024-11-17 c60d4cd
Update CHANGELOG.md
