### 2024-11-17 dc5f4f5
Add the Van-der-Waals model for JAX
