### 2024-11-17 e5e524d
Update CHANGELOG.md
