### 2024-11-17 9539873
Add the Van-der-Waals model for JAX (#902)
* Add the Van-der-Waals model for JAX

* Update CHANGELOG.md
