+++ src/felupe/constitution/__init__.py
@@ -68,6 +68,7 @@ try:
         ogden,
         ogden_roxburgh,
         saint_venant_kirchhoff,
+        saint_venant_kirchhoff_orthotropic,
         storakers,
         third_order_deformation,
         van_der_waals,
@@ -92,6 +93,7 @@ try:
         "ogden",
         "ogden_roxburgh",
         "saint_venant_kirchhoff",
+        "saint_venant_kirchhoff_orthotropic",
         "storakers",
         "third_order_deformation",
         "van_der_waals",
