+++ src/felupe/constitution/jax/models/hyperelastic/_extended_tube.py
@@ -17,22 +17,20 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 """
 from functools import wraps
 
-from jax.numpy import array, diag, sqrt
-from jax.numpy import sum as asum
-from jax.numpy.linalg import eigvalsh
+from jax.numpy import array, diag, log, sqrt, trace
+from jax.numpy.linalg import det, eigvalsh
 
-from ....tensortrax.models.hyperelastic import storakers as storakers_docstring
+from ....tensortrax.models.hyperelastic import extended_tube as extended_tube_docstring
 
 
-@wraps(storakers_docstring)
-def storakers(C, mu, alpha, beta):
-    λ1, λ2, λ3 = sqrt(eigvalsh(C + diag(array([0, -1e-4, 1e-4]))))
-    J = λ1 * λ2 * λ3
-
-    μ = array(mu)
-    α = array(alpha)
-    β = array(beta)
-
-    return asum(
-        2 * μ / α**2 * (λ1**α + λ2**α + λ3**α - 3 + (J ** (-α * β) - 1) / β)
-    )
+@wraps(extended_tube_docstring)
+def extended_tube(C, Gc, delta, Ge, beta):
+    J3 = det(C) ** (-1 / 3)
+    D = J3 * trace(C)
+    λ1, λ2, λ3 = sqrt(J3 * eigvalsh(C + diag(array([0, 1e-4, -1e-4]))))
+    β = beta
+    δ = delta
+    γ = (1 - δ**2) * (D - 3) / (1 - δ**2 * (D - 3))
+    Wc = Gc / 2 * (γ + log(1 - δ**2 * (D - 3)))
+    We = 2 * Ge / β**2 * (λ1**-β + λ2**-β + λ3**-β - 3)
+    return Wc + We
