+++ tests/test_constitution.py
@@ -342,6 +342,17 @@ def test_umat_hyperelastic():
     for model, kwargs, incompressible in [
         (neo_hooke, {"mu": 1}, True),
         (fem.saint_venant_kirchhoff, {"mu": 1, "lmbda": 20.0}, False),
+        (
+            fem.saint_venant_kirchhoff_orthotropic,
+            {
+                "mu": [1, 1, 1],
+                "lmbda": [20, 20, 20, 20, 20, 20],
+                "r1": np.eye(3)[:, 0],
+                "r2": np.eye(3)[:, 1],
+                "r3": np.eye(3)[:, 2],
+            },
+            False,
+        ),
         (fem.neo_hooke, {"mu": 1}, True),
         (fem.mooney_rivlin, {"C10": 0.3, "C01": 0.8}, True),
         (fem.yeoh, {"C10": 0.5, "C20": -0.1, "C30": 0.02}, True),
