### 2024-11-18 00e1ef7
Aktualisieren von CHANGELOG.md
