### 2024-11-18 7f08905
Add more hyperelastic models (#903)
* Add more hyperelastic models

SVK-ortho (tensortrax), extended-tube (JAX), neo-hooke (JAX)

* Aktualisieren von CHANGELOG.md

* Aktualisieren von test_constitution_jax.py
