+++ src/felupe/constitution/jax/models/hyperelastic/_van_der_waals.py
@@ -17,8 +17,7 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 """
 from functools import wraps
 
-from jax.lax import select
-from jax.numpy import isclose, log, sqrt, trace
+from jax.numpy import log, sqrt, trace
 from jax.numpy.linalg import det
 
 from ....tensortrax.models.hyperelastic import van_der_waals as van_der_waals_docstring
@@ -30,7 +29,7 @@ def van_der_waals(C, mu, limit, a, beta):
     I1 = J3 * trace(C)
     I2 = (trace(C) ** 2 - J3**2 * trace(C @ C)) / 2
     Im = (1 - beta) * I1 + beta * I2
-    Im = select(isclose(Im, 3), Im + 1e-6, Im)
+    Im += 1e-4
     eta = sqrt((Im - 3) / (limit**2 - 3))
     return mu * (
         -(limit**2 - 3) * (log(1 - eta) + eta) - 2 / 3 * a * ((Im - 3) / 2) ** (3 / 2)
