+++ src/felupe/__init__.py
@@ -263,6 +263,7 @@ try:
         ogden,
         ogden_roxburgh,
         saint_venant_kirchhoff,
+        saint_venant_kirchhoff_orthotropic,
         storakers,
         third_order_deformation,
         total_lagrange,
@@ -292,6 +293,7 @@ try:
         "ogden",
         "ogden_roxburgh",
         "saint_venant_kirchhoff",
+        "saint_venant_kirchhoff_orthotropic",
         "storakers",
         "third_order_deformation",
         "van_der_waals",
