+++ src/felupe/constitution/jax/models/hyperelastic/_blatz_ko.py
@@ -25,7 +25,6 @@ from ....tensortrax.models.hyperelastic import blatz_ko as blatz_ko_docstring
 
 @wraps(blatz_ko_docstring)
 def blatz_ko(C, mu):
-
     I1 = trace(C)
     I2 = (I1**2 - trace(C @ C)) / 2
     I3 = det(C)
