### 2024-11-18 e741f0d
Aktualisieren von _saint_venant_kirchhoff_orthotropic.py
