### 2024-11-19 7ba5532
Add `constitution.lame_converter_orthotropic()`
