### 2024-11-19 98f42f3
Add `constitution.lame_converter_orthotropic()` (#904)
