### 2024-11-20 a7aa074
Add an optional strain exponent for the SVK-material
