### 2024-11-21 90988ac
Make the third orthotropic-direction optional
in `saint_venant_kirchhoff_orthotropic()`
