### 2024-11-21 02e014b
Add an optional strain exponent for the SVK-material (#905)
* Add an optional strain exponent for the SVK-material

* Make the third orthotropic-direction optional

in `saint_venant_kirchhoff_orthotropic()`
