### 2024-11-21 fc59ddc
Update _lame_converter.py
