### 2024-11-23 452411a
Add a `paper.md` for JOSS (draft)
