+++ paper/paper.bib
@@ -1,0 +1,275 @@
+@Article{dutzler2021,
+  author    = {Dutzler, Andreas and Buzzi, Christian and Leitner, Martin},
+  journal   = {Journal of Applied Engineering Design and Simulation},
+  title     = {Nondimensional translational characteristics of elastomer components},
+  year      = {2021},
+  issn      = {2805-5756},
+  month     = sep,
+  number    = {1},
+  volume    = {1},
+  doi       = {10.24191/jaeds.v1i1.20},
+  publisher = {UiTM Press, Universiti Teknologi MARA},
+}
+
+@Article{torggler2023,
+  author    = {Torggler, J. and Dutzler, A. and Oberdorfer, B. and Faethe, T. and Müller, H. and Buzzi, C. and Leitner, M.},
+  journal   = {Applied Composite Materials},
+  title     = {Investigating Damage Mechanisms in Cord-Rubber Composite Air Spring Bellows of Rail Vehicles and Representative Specimen Design},
+  year      = {2023},
+  issn      = {1573-4897},
+  month     = aug,
+  number    = {6},
+  pages     = {1979--1999},
+  volume    = {30},
+  doi       = {10.1007/s10443-023-10157-1},
+  publisher = {Springer Science and Business Media LLC},
+}
+
+@Book{bonetwood,
+  author    = {Bonet, Javier and Wood, Richard D.},
+  publisher = {Cambridge University Press},
+  title     = {Nonlinear Continuum Mechanics for Finite Element Analysis},
+  year      = {2008},
+  isbn      = {9780511755446},
+  month     = mar,
+  doi       = {10.1017/cbo9780511755446},
+}
+
+@Article{scikitfem,
+  author    = {Gustafsson, Tom and McBain, G.},
+  journal   = {Journal of Open Source Software},
+  title     = {scikit-fem: A Python package for finite element assembly},
+  year      = {2020},
+  issn      = {2475-9066},
+  month     = aug,
+  number    = {52},
+  pages     = {2369},
+  volume    = {5},
+  doi       = {10.21105/joss.02369},
+  publisher = {The Open Journal},
+}
+
+@Book{bathe,
+  author    = {Bathe, Klaus-Jürgen},
+  publisher = {Bathe},
+  title     = {Finite element procedures},
+  year      = {2006},
+  address   = {[Boston, Mass.]},
+  isbn      = {9780979004902},
+  note      = {Frühere Ausg. u.d.T.: Bathe, Klaus-Jürgen: Finite element procedures in engineering analysis},
+  pagetotal = {1037},
+  ppn_gvk   = {641562853},
+}
+
+@Book{zienkiewicz,
+  author    = {Zienkiewicz, O. C.},
+  editor    = {Robert L. Taylor and Jianzhong Zhu and Olek C Zienkiewicz},
+  publisher = {Elsevier Science & Technology},
+  title     = {Finite Element Method},
+  year      = {2013},
+  address   = {Oxford},
+  edition   = {7th ed.},
+  isbn      = {9780080951355},
+  note      = {Description based on publisher supplied metadata and other sources.},
+  series    = {The Finite Element Method Ser},
+  pagetotal = {1753},
+  ppn_gvk   = {1658035038},
+  subtitle  = {Its Basis and Fundamentals},
+}
+
+@Article{pyvista,
+  author    = {Sullivan, C. and Kaszynski, Alexander},
+  journal   = {Journal of Open Source Software},
+  title     = {PyVista: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK)},
+  year      = {2019},
+  issn      = {2475-9066},
+  month     = may,
[…]
