+++ paper/paper.md
@@ -1,0 +1,100 @@
+---
+title: 'FElupe: Finite element analysis for continuum mechanics of solid bodies'
+tags:
+  - python
+  - finite-elements
+  - hyperelasticity
+  - computations-mechanics
+  - scientific-computing
+authors:
+  - name: Andreas Dutzler
+    orcid: 0000-0002-9383-9686
+    affiliation: [ 1, 2 ]
+  - name: Martin Leitner
+    orcid: 0000-0002-3530-1183
+    affiliation: 1
+affiliations:
+  - index: 1
+    name: Institute of Structural Durability and Railway Technology, Graz University of Technology, Austria
+  - index: 2
+    name: Siemens Mobility Austria GmbH, Austria
+date: 18 October 2024
+bibliography: paper.bib
+---
+
+# Summary
+FElupe is a Python package for finite element analysis focusing on the formulation and
+numerical solution of nonlinear problems in continuum mechanics of solid bodies. This package is intended for scientific research, but is also suitable for running nonlinear simulations in general. In addition to the transformation of general weak forms into sparse vectors and matrices, FElupe provides an efficient high-level abstraction layer for the simulation of the deformation of solid bodies. The finite element method, as used in FElupe, is generally based on the preliminary works by [@bonetwood], [@bathe] and [@zienkiewicz].
+
+## Highlights
+- pure Python package built with NumPy and SciPy
+- easy-to-learn and productive high-level API
+- nonlinear deformation of solid bodies with interactive views
+- hyperelastic material models with automatic differentiation
+
+# Statement of need
+There are well-established Python packages available for finite element analysis. These packages are either distributed as binary packages or need to be compiled on installation, like FEniCSx [@fenicsx], GetFEM [@getfem] or SfePy [@sfepy]. JAX-FEM [@jaxfem], which is built on JAX [@jax], is a pure Python package but requires many dependencies in its recommended environment. `scikit-fem` [@scikitfem] is a pure Python package with minimal dependencies but with a more general scope [@scikitfem]. FElupe is both easy-to-install as well as easy-to-use in its target domain of hyperelastic solid bodies.
+
+The performance of FElupe is good for a non-compiled package but mediocre in comparison to compiled codes. However, it is still well-suited for up to mid-sized problems, i.e. up to $10^5$ degrees of freedom, when basic hyperelastic model formulations are used. A performance benchmark for times spent on stiffness matrix assembly is included in the documentation. Internally, efficient NumPy [@numpy] based math is realized by element-wise operating trailing axes [@scikitfem]. An all-at-once approach per operation is used instead of a cell-by-cell evaluation loop. The constitutive material formulation class is backend agnostic: FElupe provides NumPy-arrays as input arguments and requires NumPy-arrays as return values. This enables backends like JAX [@jax] or PyTorch [@pytorch] to be used. Interactive views of meshes, fields and solid bodies are enabled by PyVista [@pyvista]. The capabilities of FElupe may be enhanced with additional Python packages, e.g. `meshio` [@meshio], `matadi` [@matadi], `tensortrax` [@tensortrax], `hyperelastic` [@hyperelastic] or `feplot` [@feplot].
+
+
+
+# Features
+The essential high-level parts of solving problems with FElupe include a field, a solid body, boundary conditions and a job. With the combination of a mesh, a finite element formulation and a quadrature rule, a numeric region is created. A field for a field container is further created on top of this numeric region, see \autoref{fig:field}.
+
+![Schematic representation of classes needed to create a field container.\label{fig:field}](field.pdf)
+
[…]
