### 2024-11-23 #212 @adtzlr
Hi @tkoyama010, @cmarmo!

The draft for the FElupe JOSS-paper is ready and it is included in the repo `paper/paper.md`. Meanwhile, a compiled PDF is available [here](https://github.com/adtzlr/felupe-paper/blob/main/paper/paper.pdf).
