### 2024-11-26 fa4a7d4
Add `SolidBody.evaluate.mass(density=1.0)`
