### 2024-11-26 ef42d7c
Add tests and update changelog
