+++ src/felupe/__about__.py
@@ -1,1 +1,1 @@
-__version__ = "9.1.0"
+__version__ = "9.2.0-dev"
