### 2024-11-26 9407533
Add `SolidBody.evaluate.mass(density=1.0)` (#906)
* Add `SolidBody.evaluate.mass(density=1.0)`

* Add tests and update changelog

* format black
