### 2024-11-27 ef945f3
Autodetect stress-type in `SolidBody.plot()` from string
