+++ tests/test_mechanics.py
@@ -276,6 +276,10 @@ def test_solidbody_incompressible():
         F2 = b._extract(u)
         assert np.allclose(F1, F2)
 
+        p1 = b.evaluate.stress()
+        p2 = b.evaluate.stress(u)
+        assert np.allclose(p1, p2)
+
         s1 = b.evaluate.cauchy_stress()
         s2 = b.evaluate.cauchy_stress(u)
         assert np.allclose(s1, s2)
@@ -520,12 +524,10 @@ def test_view():
     umat = fem.LinearElasticPlaneStress(E=1, nu=0.3)
 
     solid = fem.SolidBody(umat, field)
-    with pytest.warns():
-        plotter = solid.plot(off_screen=True)
+    plotter = solid.plot(off_screen=True)
 
     solid = fem.SolidBodyNearlyIncompressible(umat, field, bulk=0)
-    with pytest.warns():
-        plotter = solid.plot(off_screen=True)
+    plotter = solid.plot(off_screen=True)
 
 
 def test_threefield():
