### 2024-11-27 a8b5740
Update _solidbody.py
