### 2024-11-27 ddae2d7
Update test_mechanics.py
