### 2024-11-27 693fa5d
Update _solidbody.py
