### 2024-11-27 3806bc4
Add `FreeVibration(items, boundaries)`
with methods to `evaluate()` and `extract()` the n-th eigenvector to a field
