+++ src/felupe/mechanics/_solidbody_incompressible.py
@@ -56,6 +56,8 @@ class SolidBodyNearlyIncompressible(Solid):
         A valid initial state for a (nearly) incompressible solid (default is None).
     statevars : ndarray or None, optional
         Array of initial internal state variables (default is None).
+    density : float or None, optional
+        The density of the solid body.
 
     Notes
     -----
@@ -315,10 +317,11 @@ class SolidBodyNearlyIncompressible(Solid):
         for (nearly) incompressible solid bodies.
     """
 
-    def __init__(self, umat, field, bulk, state=None, statevars=None):
+    def __init__(self, umat, field, bulk, state=None, statevars=None, density=None):
         self.umat = umat
         self.field = field
         self.bulk = bulk
+        self.density = density
 
         self._area_change = AreaChange()
         self._form = IntegralForm
@@ -540,9 +543,13 @@ class SolidBodyNearlyIncompressible(Solid):
 
         return dot(P, transpose(F)) / J
 
-    def _mass(self, density=1.0):
+    def _mass(self, density=None):
+        if density is None:
+            density = self.density
+
         field = self.field[0].as_container()
         dim = field[0].dim
+
         form = self._form(
             fun=[density * np.eye(dim).reshape(dim, dim, 1, 1)],
             v=field,
@@ -551,4 +558,5 @@ class SolidBodyNearlyIncompressible(Solid):
             grad_v=[False],
             grad_u=[False],
         )
+
         return form.assemble()
