### 2024-11-27 e2305db
Add `SolidBody.evaluate.stress()` and autodetect the stress-type in `SolidBody.plot(name)` from `name` (#908)
* Autodetect stress-type in `SolidBody.plot()` from string

* Update CHANGELOG.md

* Update _solidbody.py

* Update test_readme.py

* Update test_mechanics.py

* Update _solidbody.py
