### 2024-11-28 94d0b8a
Update _free_vibration.py
