### 2024-11-28 0939857
Create test_free_vibration.py
