+++ tests/test_free_vibration.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 import felupe as fem
 
 
