+++ src/felupe/mechanics/_free_vibration.py
@@ -19,6 +19,7 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 import numpy as np
 from scipy.sparse import csr_matrix
 from scipy.sparse.linalg import eigsh
+
 from ..dof import partition
 
 
