### 2024-11-28 bae3834
Update test_mechanics.py
