+++ tests/test_mechanics.py
@@ -235,7 +235,7 @@ def test_solidbody_incompressible():
 
     umat = fem.OgdenRoxburgh(fem.NeoHooke(mu=1), r=3, m=1, beta=0)
     b = fem.SolidBodyNearlyIncompressible(
-        umat=umat, field=u, bulk=5000, state=fem.StateNearlyIncompressible(u)
+        umat=umat, field=u, bulk=5000, state=fem.StateNearlyIncompressible(u), density=1.0
     )
 
     M = b.assemble.mass()
