### 2024-11-28 5692cac
Update test_mechanics.py
