### 2024-11-28 d5b28a2
add test, format black
