### 2024-11-28 3ce9893
Update test_free_vibration.py
