### 2024-11-28 c77b69f
Add `FreeVibration(items, boundaries)` (#909)
* Add `FreeVibration(items, boundaries)`

with methods to `evaluate()` and `extract()` the n-th eigenvector to a field

* Update _free_vibration.py

* Create test_free_vibration.py

* add test

* Update test_mechanics.py

* Update test_mechanics.py

* add test, format black

* Update test_free_vibration.py

* update the free-vibration test

* Update mechanics.rst
