### 2024-11-28 444a12f
Enhance `hello_world()` with new arguments
