### 2024-11-30 43a7705
`paper.md` (draft): simplify the statement on performance
