### 2024-11-30 73acbc9
Enhance `Boundary`: Support multiaxial prescribed values
and broadcast them (if necessary)
