### 2024-11-30 088e64f
Enhance `Boundary`: Add support for multiaxial prescribed `value`-argument (#912)
* Enhance `Boundary`: Support multiaxial prescribed values

and broadcast them (if necessary)

* format black

* Update CHANGELOG.md

* Update test_dof.py
