### 2024-11-30 1b30023
Examples: Simplify and remove `ravel()` from nd-prescribed `values`
this is now handled automatically
