### 2024-12-01 6c936b8
Add `Boundary.plot()`
