### 2024-12-01 6b59077
Add `Boundary.plot()` and fix `Boundary(..., mode="and")` (#913)
* Add `Boundary.plot()`

* Update CHANGELOG.md

* Update extut03_building_blocks.py
