+++ src/felupe/dof/_boundary.py
@@ -92,18 +92,10 @@ class Boundary:
         :context:
         :force_static:
 
-        >>> import pyvista as pv
-        >>>
         >>> right = fem.Boundary(displacement, fx=x.max())
         >>> right = fem.Boundary(displacement, fx=lambda x: np.isclose(x, x.max()))
         >>>
-        >>> plotter = pv.Plotter()
-        >>> actor = plotter.add_points(
-        ...     mesh.points[right.points],
-        ...     point_size=20,
-        ...     color="red",
-        ... )
-        >>> mesh.plot(plotter=plotter, opacity=0.7).show()
+        >>> right.plot(color="red", plotter=mesh.plot(opacity=0.7)).show()
 
     If ``fx`` and ``fy`` are given, the masks are combined by *logical-or*.
 
@@ -112,14 +104,7 @@ class Boundary:
         :force_static:
 
         >>> axes = fem.Boundary(displacement, fx=0, fy=0, mode="or")
-        >>>
-        >>> plotter = pv.Plotter()
-        >>> actor = plotter.add_points(
-        ...     mesh.points[axes.points],
-        ...     point_size=20,
-        ...     color="red",
-        ... )
-        >>> mesh.plot(plotter=plotter, opacity=0.7).show()
+        >>> axes.plot(color="red", plotter=mesh.plot(opacity=0.7)).show()
 
     This may be changed to *logical-and* if desired.
 
@@ -128,14 +113,7 @@ class Boundary:
         :force_static:
 
         >>> center = fem.Boundary(displacement, fx=0, fy=0, mode="and")
-        >>>
-        >>> plotter = pv.Plotter()
-        >>> actor = plotter.add_points(
-        ...     mesh.points[center.points],
-        ...     point_size=20,
-        ...     color="red",
-        ... )
-        >>> mesh.plot(plotter=plotter, opacity=0.7).show()
+        >>> center.plot(color="red", plotter=mesh.plot(opacity=0.7)).show()
 
     For the most-general case, a user-defined boolean mask for the selection of the
     mesh-points is provided. While the two upper methods are useful to select
@@ -149,13 +127,7 @@ class Boundary:
         >>> mask = np.logical_and(np.isclose(x**2 + y**2, 1), x >= 0)
         >>> surface = fem.Boundary(displacement, mask=mask)
         >>>
-        >>> plotter = pv.Plotter()
-        >>> actor = plotter.add_points(
-        ...     mesh.points[surface.points],
-        ...     point_size=20,
-        ...     color="red",
-        ... )
-        >>> mesh.plot(plotter=plotter, opacity=0.7).show()
+        >>> surface.plot(color="red", plotter=mesh.plot(opacity=0.7)).show()
 
     The application of a new mask allows to change the selected points of an existing
     boundary condition.
@@ -167,13 +139,7 @@ class Boundary:
         >>> new_mask = np.logical_and(mask, y <= 0)
         >>> surface.apply_mask(new_mask)
         >>>
-        >>> plotter = pv.Plotter()
-        >>> actor = plotter.add_points(
-        ...     mesh.points[surface.points],
-        ...     point_size=20,
-        ...     color="red",
-        ... )
-        >>> mesh.plot(plotter=plotter, opacity=0.7).show()
+        >>> surface.plot(color="red", plotter=mesh.plot(opacity=0.7)).show()
 
     A boundary condition may be skipped on given axes, i.e. if only the x-components
     of a field should be prescribed on the selected points, then the y-axis must
@@ -184,14 +150,7 @@ class Boundary:
         :force_static:
 
         >>> axes_x = fem.Boundary(displacement, fx=0, fy=0, skip=(False, True))
-        >>>
[…]
