### 2024-12-01 389ed1d
Add `BoundaryDict`
 with methods to plot, screenshot and imshow
