+++ tests/test_dof.py
@@ -182,13 +182,13 @@ def test_boundary_plot():
 
 def test_boundary_dict():
     field = fem.Field(
-        fem.RegionHexahedron(fem.Cube(b=(3, 1, 1), n=5)), dim=1
+        fem.RegionHexahedron(fem.Cube(b=(3, 1, 1), n=5)), dim=3
     ).as_container()
     boundaries = fem.BoundaryDict(
         left=fem.Boundary(field[0], fx=0, skip=(0, 0, 1)),
         right=fem.Boundary(field[0], name="my_label", fx=3, skip=(0, 1, 0)),
     )
-    # plotter = boundaries.plot()
+    plotter = boundaries.plot()
     # img = boundaries.screenshot()
     # ax = boundaries.imshow()
 
