### 2024-12-01 5e627cc
Ensure 3d-data in `Boundary.plot()` for PyVista
