+++ src/felupe/dof/_boundary.py
@@ -293,21 +293,22 @@ class Boundary:
             if label is None:
                 label = self.name
 
+            points = np.pad(mesh.points, ((0, 0), (0, 3 - mesh.dim)))
             magnitude = min(mesh.points.max(axis=0) - mesh.points.min(axis=0)) * scale
 
             _ = plotter.add_points(
-                mesh.points[self.points],
+                points[self.points],
                 color=color,
                 point_size=point_size,
                 label=label,
             )
 
-            for skip, direction in zip(self.skip, np.eye(mesh.dim)):
+            for skip, direction in zip(self.skip, np.eye(3)):
                 if not skip:
-                    end = mesh.points[self.points] + direction * magnitude
+                    end = points[self.points] + direction * magnitude
                     _ = plotter.add_lines(
-                        np.hstack([mesh.points[self.points], end]).reshape(
-                            -1, mesh.dim
+                        np.hstack([points[self.points], end]).reshape(
+                            -1, 3
                         ),
                         color=color,
                         width=width,
