### 2024-12-02 11d5233
Add `Boundary.plot(..., show_points=True, show_lines=True)`
