### 2024-12-02 6c0afa2
Add `BoundaryDict` (#915)
* Add `BoundaryDict`

 with methods to plot, screenshot and imshow

* Update test_dof.py

* Update test_dof.py

* Ensure 3d-data in `Boundary.plot()` for PyVista

* Update test_dof.py

* Add `Boundary.plot(..., show_points=True, show_lines=True)`
