+++ src/felupe/dof/__init__.py
@@ -1,9 +1,11 @@
 from ._boundary import Boundary
+from ._dict import BoundaryDict
 from ._loadcase import biaxial, shear, symmetry, uniaxial
 from ._tools import apply, get_dof0, get_dof1, partition
 
 __all__ = [
     "Boundary",
+    "BoundaryDict",
     "biaxial",
     "shear",
     "symmetry",
