### 2024-12-02 139383d
Update _loadcase.py
