### 2024-12-03 2c0396d
Add `PointLoad.plot()`
