+++ tests/test_mechanics.py
@@ -504,6 +504,7 @@ def test_load():
 
         body = fem.SolidBody(umat, field)
         load = fem.PointLoad(field, mask, values=values, axisymmetric=axi)
+        plotter = load.plot(plotter=field.region.mesh.plot(off_screen=True))
 
         bounds = {"fix": fem.Boundary(field[0], fx=lambda x: x == 0)}
         dof0, dof1 = fem.dof.partition(field, bounds)
