### 2024-12-04 2680088
Update test_mechanics.py
