### 2024-12-04 ada2a2f
Add `PointLoad.plot()` (#916)
* Add `PointLoad.plot()`

* Update _pointload.py

* Update test_mechanics.py
