### 2024-12-04 #212 @nicoguaro
Hello @cmarmo, I think I can review this submission, but could you explain the process to me? Does it follow JOSS guidelines?
