### 2024-12-04 #212 @tkoyama010
> Hello @cmarmo, I think I can review this submission, but could you explain the process to me? Does it follow JOSS guidelines?

@nicoguaro
Thank you for contacting We will be happy to assign you as a reviewer. You might want to have a look to [our review guide](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html) to have a better idea of our review process.
Also feel free to ask me any question if you need more details.
