### 2024-12-07 #212 @nicoguaro
@tkoyama010, I think I have got it after reading the guide. If I have some question, I can ask them here.

I will start my review but it will take me some time since we are closing the semester and have a master thesis to review in my queue.
