### 2024-12-08 d80bcee
Enhance `math.linsteps(..., values=0.0)`
