### 2024-12-08 b2e6242
Use `BoundaryDict` in examples
