### 2024-12-08 be2f25c
Enhance `math.linsteps(..., values=0.0)` (#917)
* Enhance `math.linsteps(..., values=0.0)`

* `BoundaryDict().plot()` pass missing `size` arg

* Use `BoundaryDict` in examples
