### 2024-12-10 e28bf5b
Enhance `FieldContainer.link()` (#918)
