### 2024-12-11 8607b63
Update _tensor.py
