### 2024-12-17 bfba653
Add `SolidBody(..., block=True, apply=None)`
to apply a final callable in `SolidBody.assemble.vector()` or `SolidBody.assemble.matrix()`.
