### 2024-12-17 326a8eb
Add `SolidBody(..., multiplier=None)`
