### 2024-12-17 a518200
Update test_mechanics.py
