### 2024-12-17 a666ecc
Add `SolidBody(..., block=True, apply=None)` (#920)
* Add `SolidBody(..., block=True, apply=None)`

to apply a final callable in `SolidBody.assemble.vector()` or `SolidBody.assemble.matrix()`.

* Add `SolidBody(..., multiplier=None)`

* Update test_mechanics.py
