### 2024-12-18 15ccae9
Enhance `IntegralForm`: Non-symmetric mixed forms and `None`
