### 2024-12-18 67e10df
Enhance `IntegralForm`: Non-symmetric mixed forms and `None` (#922)
* Enhance `IntegralForm`: Non-symmetric mixed forms and `None`

* Update test_mechanics.py

* Add more tests
