### 2025-01-02 4aaf105
Fix Ex20: Merge points of top-level mesh container
also rename `gamma` to `gamma0` and add the parameter `kr`
