+++ src/felupe/element/_triangle.py
@@ -58,7 +58,6 @@ class Triangle(Element):
     """
 
     def __init__(self):
-        super().__init__(shape=(3, 2))
         self.points = np.array([[0, 0], [1, 0], [0, 1]], dtype=float)
         self.cells = np.arange(len(self.points)).reshape(1, -1)
         self.cell_type = "triangle"
@@ -115,7 +114,6 @@ class TriangleMINI(Element):
     """
 
     def __init__(self, bubble_multiplier=1.0):
-        super().__init__(shape=(4, 2))
         self.points = np.array([[0, 0], [1, 0], [0, 1], [1 / 3, 1 / 3]], dtype=float)
         self.cells = np.arange(len(self.points) - 1).reshape(1, -1)
         self.cell_type = "triangle"
@@ -198,7 +196,6 @@ class QuadraticTriangle(Element):
     """
 
     def __init__(self):
-        super().__init__(shape=(6, 2))
         self.points = np.zeros(self.shape)
         self.points[:3] = np.array([[0, 0], [1, 0], [0, 1]], dtype=float)
         self.points[3] = np.mean(self.points[[0, 1]], axis=0)
