### 2025-01-03 235a641
Update _lagrange.py
