### 2025-01-03 3d7944d
Don't use `Element.shape` in `"tetra10"` and `"triangle6"`
