### 2025-01-03 2ab6697
Simplify init of base-class `element.Element` (#924)
* Simplify `Element`

and add `element.Element` to the top-level namespace

* Update _lagrange.py

* Don't use `Element.shape` in `"tetra10"` and `"triangle6"`
