+++ src/felupe/__init__.py
@@ -46,6 +46,7 @@ from .element import (
     BiQuadraticQuad,
     ConstantHexahedron,
     ConstantQuad,
+    Element,
     Hexahedron,
     Line,
     Quad,
@@ -173,6 +174,7 @@ __all__ = [
     "BiQuadraticQuad",
     "ConstantHexahedron",
     "ConstantQuad",
+    "Element",
     "Hexahedron",
     "Line",
     "Quad",
