### 2025-01-03 1132065
Add a docstring for `Element`
