### 2025-01-22 ea697b8
Fix `hello_word(planestrain=True)` (#928)
