### 2025-01-23 586cab1
Revert "Update .readthedocs.yaml"
This reverts commit c5d0e2a083d82ae19207bf64af5b22ed788538cb.
