+++ .readthedocs.yaml
@@ -3,23 +3,21 @@ version: 2
 build:
   os: ubuntu-lts-latest
   tools:
-    python: "3.13"
+    python: "3.12"
   apt_packages:
     - libgl1-mesa-dev
     - xvfb
     - pandoc
-  jobs:
-    create_environment:
-      - asdf plugin add uv
-      - asdf install uv latest
-      - asdf global uv latest
-      - uv venv
-    install:
-      - uv pip install .[all,docs,examples]
-    build:
-      html:
-        - uv run sphinx-build -T -b html docs $READTHEDOCS_OUTPUT/html
 
 sphinx:
   configuration: docs/conf.py
   fail_on_warning: false
+
+python:
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - all
+        - docs
+        - examples
