### 2025-01-23 ee0f153
Add more badges
repostatus, supported Python versions
