### 2025-02-06 #212 @nicoguaro
@adtzlr, I ran the first round of reviews on installing the software, reviewing the documentation, and running some examples. The next homework in the nonlinear FEM course I am lecturing involves some hyperelastic simulations, so I will try to run them in Felupe.
