### 2025-02-06 adtzlr/felupe#931 @adtzlr
Hi @nicoguaro!

- [x] `SolidBodyGravity`: I keep the deprecated `SolidBodyGravity` in the tests, just to be sure its tested until it is removed. The deprecation warning is okay.
- [x] The negative volumes for cells (Jacobians) are also part of the test suite - I want to be sure that the warning is raised. This happens quite frequently with imported meshes from different software packages via meshio.
- [ ] I'll have a look on the math tests. I think they are okay - some hyperelastic models are not well-defined in the undeformed state...
