### 2025-02-07 7f40c66
Fix `RegionQuadraticQuadBoundary` (#933)
* Fix `RegionQuadraticQuadBoundary`

fix the cells-array for midpoints of edges

* Capture some warnings in the tests

* Optionally close figures in tests
