+++ tests/test_job.py
@@ -55,7 +55,9 @@ def test_job():
 
 
 def test_job_xdmf():
-    field, step = pre(SolidBodyForce=fem.SolidBodyGravity)
+    with pytest.warns():  # solidbodygravity is deprecated
+        field, step = pre(SolidBodyForce=fem.SolidBodyGravity)
+        
     job = fem.Job(steps=[step])
     job.evaluate()
 
