### 2025-02-08 e855101
Don't create a figure if error in `umat.plot()`
