### 2025-02-08 adtzlr/felupe#931 @adtzlr
@nicoguaro, how would you like to handle the pyOpenSci-related Issues? Should I keep them open? Should I close them when I think they are finished? Would you like to close them once the fixes / enhancements are okay for you?
