### 2025-02-08 adtzlr/felupe#931 @adtzlr
Okay, @nicoguaro, then let's close this issue, all warnings are captured in in the tests now. One bug in boundary-regions for quadratic quads was detected in the warnings and is fixed now.
