### 2025-02-10 0b3a637
Add docstrings to `solve.solve()` and `solve.partition()`
