+++ src/felupe/solve/_solve.py
@@ -25,7 +25,67 @@ from ..math import values
 def partition(v, K, dof1, dof0, r=None):
     """Perform partitioning of field values (unknowns), (stiffness) matrix
     and (residuals) vector with given lists of active (dof1) and
-    prescribed degrees of freedom (dof0)."""
+    prescribed degrees of freedom (dof0).
+
+    Parameters
+    ----------
+    v : FieldContainer
+        A field container with the fields. It is used to extract and concatenate the
+        1d-array of unknows.
+    K : scipy.sparse.spmatrix
+        A two-dimensional sparse (stiffness) matrix.
+    dof1 : list of int
+        List of active degrees of freedom (zero-indexed).
+    dof0 : list of int
+        List of prescribed degrees of freedom (zero-indexed).
+    r : 1d-array or None, optional
+        The 1d-vector of residuals. If None, the residuals are set to zero. Default is
+        None.
+
+    Returns
+    -------
+    u : 1d-array
+        The concatenated full 1d-array of unknowns, extracted from the field container.
+    u0 : 1d-array
+        The prescribed unknowns.
+    K11 : scipy.sparse.spmatrix
+        The active-active block of the two-dimensional sparse (stiffness) matrix.
+    K10 : scipy.sparse.spmatrix
+        The active-prescribed block of the two-dimensional sparse (stiffness) matrix.
+    dof1 : list of int
+        List of active degrees of freedom (zero-indexed).
+    dof0 : list of int
+        List of prescribed degrees of freedom (zero-indexed).
+    r1 : 1d-array
+        The 1d-vector of active residuals.
+
+    Examples
+    --------
+    ..  pyvista-plot::
+        :force_static:
+
+        >>> import felupe as fem
+        >>>
+        >>> mesh = fem.Rectangle(n=3)
+        >>> region = fem.RegionQuad(mesh)
+        >>> field = fem.FieldPlaneStrain(region, dim=2).as_container()
+        >>> boundaries, loadcase = fem.dof.uniaxial(field, clamped=True)
+        >>> umat = fem.NeoHooke(mu=1.0, bulk=2.0)
+        >>> solid = fem.SolidBody(umat, field)
+        >>>
+        >>> K = solid.assemble.matrix()
+        >>> r = solid.assemble.vector()
+        >>>
+        >>> dof0 = loadcase["dof0"]
+        >>> dof1 = loadcase["dof1"]
+        >>> ext0 = loadcase["ext0"]
+        >>>
+        >>> system = fem.solve.partition(field, K, dof1, dof0, r)
+
+    See Also
+    --------
+    felupe.solve.solve : Linear solution of equation system.
+    """
 
     # extract values
     u = values(v)
@@ -48,11 +108,46 @@ def partition(v, K, dof1, dof0, r=None):
 
 
 def solve(u, u0, K11, K10, dof1, dof0, r1=None, ext0=None, solver=spsolve):
-    """Linear solution of equation system with optional given values of
+    r"""Linear solution of equation system with optional given values of
     unknowns at prescribed deegrees of freedom.
 
-        K_11 du_1 = -r1 - K10 (u0_ext - u0)
+    Notes
+    -----
 
+    ..  math::
+
+        \boldsymbol{K}_{11}\ d\boldsymbol{u}_1 =
+            -\boldsymbol{r}_1 - \boldsymbol{K}_{10} (
+                \boldsymbol{u}_{0,\text{ext}} - \boldsymbol{u}_0
+            )
+
+    Examples
+    --------
+    ..  pyvista-plot::
+        :force_static:
+
+        >>> import felupe as fem
+        >>>
[…]
