+++ src/felupe/mechanics/_solidbody.py
@@ -221,6 +221,7 @@ class SolidBody(Solid):
     Examples
     --------
     ..  pyvista-plot::
+        :force_static:
 
         >>> import felupe as fem
         >>>
