### 2025-02-12 #212 @adtzlr
@nicoguaro, I think it would be a good idea to track the completed issues also here (checked items are finished).

- [x] https://github.com/adtzlr/felupe/issues/931
- [x] https://github.com/adtzlr/felupe/issues/991

@crhea93, I'll also collect your issues in this comment.
- [x] https://github.com/adtzlr/felupe/issues/953
