### 2025-02-15 2efe19a
Raise a TypeError if a container is used in a region
