### 2025-02-15 46b5093
Raise a TypeError if a container is used in a region (#939)
