### 2025-02-15 76eba51
Add `MeshContainer.as_vertex_mesh()`
