+++ src/felupe/mesh/_container.py
@@ -281,8 +281,8 @@ class MeshContainer:
 
     def as_vertex_mesh(self):
         "Return a merged vertex-mesh."
-        cells = np.unique([mesh.cells.ravel() for mesh in self.meshes]).reshape(-1, 1)
-        return Mesh(self.points, cells, cell_type=None)
+        cells = np.unique(np.concatenate([mesh.cells.ravel() for mesh in self.meshes]))
+        return Mesh(self.points, cells.reshape(-1, 1), cell_type=None)
 
     def copy(self):
         "Return a deepcopy of the mesh container."
