### 2025-02-16 6bbf048
Add `Field.from_mesh_container(container)`
to create a top-level vertex-mesh.
