### 2025-02-16 f1c9bd4
Set the cell-type to `"vertex"` in `MeshContainer.as_vertex_mesh()`
