### 2025-02-16 f061729
Add `MeshContainer.as_vertex_mesh()` and `Field.from_mesh_container()` (#940)
* Add `MeshContainer.as_vertex_mesh()`

* Fix unique cells

* Update __init__.py

* Update test_region.py

* Update test_element.py

* Update CHANGELOG.md

* Add `Field.from_mesh_container(container)`

to create a top-level vertex-mesh.

* Update CHANGELOG.md

* Add `"vertex"` celltype for PyVista conversion

* Set the cell-type to `"vertex"` in `MeshContainer.as_vertex_mesh()`

* Update test_field.py

* Update test_field.py
