### 2025-02-16 e252eaa
Update _hello_world.py
with the new `Field.from_mesh_container()`
