+++ src/felupe/tools/_hello_world.py
@@ -136,10 +136,7 @@ def hello_world(
             f"    fem.{mesh}(n=6).translate(1, axis=0),",
             "]",
             "container = fem.MeshContainer(meshes, merge=True)",
-            "mesh = container.stack()",
-            "",
-            f"region = fem.{region}(mesh)",
-            f"field = fem.FieldContainer([fem.{field}(region, dim={dim})])",
+            "field = fem.Field.from_mesh_container(container).as_container()",
             "",
             "regions = [",
             f"    fem.{region}(container.meshes[0]),",
