### 2025-02-17 3fb4c48
Add `Vertex` element and region to the API docs
