### 2025-02-17 0f93bc5
Update composite.rst
