### 2025-02-17 f3614c1
Fix `FieldContainer.plot()` for `RegionVertex` (no gradient available)
