### 2025-02-17 bfc38f4
Update composite.rst
