### 2025-02-17 f43d34a
Update composite.rst
