+++ docs/howto/composite.rst
@@ -1,5 +1,5 @@
-Composite Regions with Solid Bodies
------------------------------------
+Multiple Solid Bodies
+---------------------
 
 This section demonstrates how to set up a problem with two regions, each associated to a
 seperated solid body. Different element formulations are used for the solid bodies.
