### 2025-02-17 a864a7d
Add `GaussLobatto` quadrature
