+++ src/felupe/quadrature/__init__.py
@@ -1,4 +1,5 @@
-from ._gausslegendre import GaussLegendre, GaussLegendreBoundary
+from ._gauss_legendre import GaussLegendre, GaussLegendreBoundary
+from ._gauss_lobatto import GaussLobatto, GaussLobattoBoundary
 from ._scheme import Scheme
 from ._sphere import BazantOh
 from ._tetra import Tetrahedron
@@ -8,6 +9,8 @@ __all__ = [
     "Scheme",
     "GaussLegendre",
     "GaussLegendreBoundary",
+    "GaussLobatto",
+    "GaussLobattoBoundary",
     "Tetrahedron",
     "Triangle",
     "BazantOh",
