### 2025-02-17 81fc87a
Update test_quadrature.py
