### 2025-02-17 b37a2c3
Update test_quadrature.py
