### 2025-02-17 d36b1f3
Update _gauss_lobatto.py
