### 2025-02-17 c5fd9e9
Update _gauss_lobatto.py
