### 2025-02-17 6aa973c
Add `GaussLobatto` quadrature scheme (#943)
* Add `GaussLobatto` quadrature

* Update test_quadrature.py

* Update test_quadrature.py

* Update _gauss_lobatto.py
