### 2025-02-18 7791771
Enhance docs/howto source formatting
