+++ docs/howto/axi.rst
@@ -1,24 +1,43 @@
-Axisymmetric Problems
----------------------
+Two-dimensional Problems
+------------------------
 
-For axisymmetric analyses an axisymmetric vector-valued field has to be created for the in-plane displacements.
+For plane-strain and axisymmetric problems a vector-valued field has to be created for
+the two-dimensional in-plane displacement components.
+
+..  tab:: Axisymmetric
+
+    ..  code-block:: python
+
+        import felupe as fem
+
+        mesh = fem.Rectangle(n=3)
+        region = fem.RegionQuad(mesh)
+        displacement = fem.FieldAxisymmetric(region, dim=2)
+
+..  tab:: Plane-Strain
+
+    ..  code-block:: python
+
+        import felupe as fem
+
+        mesh = fem.Rectangle(n=3)
+        region = fem.RegionQuad(mesh)
+        displacement = fem.FieldPlaneStrain(region, dim=2)
+        
+
+The 3x3 deformation gradient for axisymmetric and plane-strain two-dimensional problems
+is obtained by the :meth:`~felupe.FieldAxisymmetric.grad` or
+:meth:`~felupe.FieldAxisymmetric.extract` methods (same for
+:class:`~felupe.FieldPlaneStrain`). For these two-dimensional fields the gradient is
+modified to return a three-dimensional gradient.
 
 ..  code-block:: python
 
-    import felupe as fem
-
-    mesh = fem.Rectangle(n=3)
-    region = fem.RegionQuad(mesh)
-    u = fem.FieldAxisymmetric(region, dim=2)
-    field = fem.FieldContainer([u])
-
-Now it gets important: The 3x3 deformation gradient for an axisymmetric problem is obtained with :meth:`~felupe.FieldAxisymmetric.grad` or :meth:`~felupe.FieldAxisymmetric.extract` methods. For a two-dimensional :class:`~felupe.FieldAxisymmetric` the gradient is modified to return a three-dimensional gradient.
-
-..  code-block:: python
-
+    field = fem.FieldContainer([displacement])
     F = field.extract(grad=True, sym=False, add_identity=True)
 
-For simplicity, let's use the isotropic hyperelastic :class:`~felupe.NeoHooke` material model formulation.
+For simplicity, let's use the isotropic hyperelastic :class:`~felupe.NeoHooke` material
+model formulation.
 
 ..  code-block:: python
 
@@ -26,7 +45,14 @@ For simplicity, let's use the isotropic hyperelastic :class:`~felupe.NeoHooke` m
 
 ..  note::
 
-    Internally, FElupe provides an adopted low-level :class:`~felupe.assembly.IntegralFormAxisymmetric` class for the integration and the sparse matrix assemblage of axisymmetric problems. It uses the additional information (e.g. radial coordinates at integration points) stored in :class:`~felupe.FieldAxisymmetric` to provide a consistent interface in comparison to :class:`~felupe.assembly.IntegralFormCartesian`. The top-level :class:`~felupe.IntegralForm` chooses the appropriate low-level integral form based on the kind of field inside the field container.
+    Internally, FElupe provides an adopted low-level
+    :class:`~felupe.assembly.IntegralFormAxisymmetric` class for the integration and the
+    sparse matrix assemblage of axisymmetric problems. It uses the additional
+    information (e.g. radial coordinates at integration points) stored in
+    :class:`~felupe.FieldAxisymmetric` to provide a consistent interface in comparison
+    to :class:`~felupe.assembly.IntegralFormCartesian`. The top-level
+    :class:`~felupe.IntegralForm` chooses the appropriate low-level integral form based
+    on the kind of field inside the field container.
 
[…]
