### 2025-02-18 2d829f1
Docs: Make more figures static
