### 2025-03-03 423808d
Enhance `mesh.interpolate_line(axis=None)` (#945)
use a curve-progress variable if `axis=None` by default
