+++ src/felupe/mesh/_interpolate.py
@@ -19,7 +19,7 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 import numpy as np
 
 
-def interpolate_line(mesh, xi, axis, Interpolator=None, **kwargs):
+def interpolate_line(mesh, xi, axis=None, Interpolator=None, **kwargs):
     r"""Return an interpolated line mesh from an existing line mesh with provided
     interpolation points on a given axis.
 
@@ -28,13 +28,15 @@ def interpolate_line(mesh, xi, axis, Interpolator=None, **kwargs):
     mesh : Mesh
         A :class:`~felupe.Mesh` with cell-type ``line``.
     xi : ndarray
-        Points at which to interpolate data.
-    axis : int
-        The axis of the points at which the data will be interpolated.
+        Points at which to interpolate data. If axis is None, a normalized curve-
+        progress must be provided (:math:`0 \le \xi \le 1`).
+    axis : int or None, optional
+        The axis of the points at which the data will be interpolated. If None, a
+        normalized curve-progress is used. Default is None.
     Interpolator : callable or None, optional
         An interpolator class (default is :class:`scipy.interpolate.PchipInterpolator`).
     **kwargs : dict, optional
-        Optional keyword arguments are passed to the interpolator.
+        Optional keyword arguments are passed to the Interpolator.
 
     Returns
     -------
@@ -51,53 +53,70 @@ def interpolate_line(mesh, xi, axis, Interpolator=None, **kwargs):
 
         >>> import felupe as fem
         >>> import numpy as np
+        >>> from scipy.interpolate import CubicSpline
         >>>
-        >>> mesh = fem.mesh.Line(n=5).expand(n=1)
+        >>> mesh = fem.mesh.Line(b=1.0, n=5).expand(n=1)
         >>> t = mesh.x.copy()
-        >>> mesh.points[:, 0] = np.sin(np.pi / 2 * t)
-        >>> mesh.points[:, 1] = np.cos(np.pi / 2 * t)
+        >>> mesh.points[:, 0] = np.sin(2 * np.pi * t)
+        >>> mesh.points[:, 1] = np.cos(2 * np.pi * t)
         >>>
-        >>> mesh.plot(style="points", color="black").show()
+        >>> mesh_new = fem.mesh.interpolate_line(
+        ...     mesh, xi=np.linspace(0, 1), Interpolator=CubicSpline, bc_type="periodic"
+        ... )
+        >>>
+        >>> mesh_new.plot(
+        ...     plotter=mesh.plot(style="points", color="red", point_size=15),
+        ...     color="black",
+        ... ).show()
 
-    ..  pyvista-plot::
-        :context:
-        :force_static:
-
-        >>> mesh_new = fem.mesh.interpolate_line(mesh, xi=np.linspace(0, 1), axis=1)
-        >>> mesh_new.plot(style="points", color="black").show()
     """
 
-    # line connectivity
-    # concatenate the first point of each cell and the last point of the last cell
-    line = np.concatenate([mesh.cells[:, :-1].ravel(), mesh.cells[-1, -1:]])
-
-    # independent spline variable
-    points = mesh.points[line, axis]
-    ascending = np.argsort(points)
-
-    # dependent spline variable(s)
-    mask = np.ones(mesh.dim, dtype=bool)
-    mask[axis] = False
-
-    axes = np.arange(mesh.dim)
-    values = mesh.points[line.reshape(-1, 1), axes[mask]]
-
-    # spline interpolator
     if Interpolator is None:
         from scipy.interpolate import PchipInterpolator as Interpolator
 
-    # create a spline
[…]
