+++ .github/workflows/upload-codecov.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.12"]
+        python-version: ["3.9", "3.13"]
     steps:
     - uses: actions/setup-python@v5
       with:
@@ -25,7 +25,7 @@ jobs:
         pip install tox
         tox -- --cov felupe --cov-report xml --cov-report term
     - name: Upload coverage to Codecov
-      if: ${{ matrix.python-version == '3.12' }}
+      if: ${{ matrix.python-version == '3.13' }}
       uses: codecov/codecov-action@v4
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
