### 2025-03-04 eb7fb90
Set version tag to 9.2.0 (#946)
* Add Python 3.13

* set version tag to 9.2.0
