+++ src/felupe/__about__.py
@@ -1,1 +1,1 @@
-__version__ = "9.2.0-dev"
+__version__ = "9.2.0"
