### 2025-03-07 4bf94dc
Fix a typo in `paper.md` (tags)
