### 2025-03-10 c608201
Add new example in docstring of `Element`
for automatic differentiation
