### 2025-04-14 #212 @nicoguaro
> > [@adtzlr](https://github.com/adtzlr), I ran the first round of reviews on installing the software, reviewing the documentation, and running some examples. The next homework in the nonlinear FEM course I am lecturing involves some hyperelastic simulations, so I will try to run them in Felupe.
>
> How did the examples run in FElupe [@nicoguaro](https://github.com/nicoguaro) ? Were you able to solve them?

I am super late on the reviews this semester. Regarding the examples for the course, I overestimated how fast we could move with the students.
