### 2025-04-28 #212 @crhea93
- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review). ✅ 2025-04-14

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README. ✅ 2025-04-14
- [x] **Installation instructions:** for the development version of the package and any non-standard dependencies in README. ✅ 2025-04-14
- [x] **Vignette(s)** demonstrating major functionality that runs successfully locally. ✅ 2025-04-28
	- The main getting started example worked flawlessly. However, I was only given the Principal values of Cauchy Stress plot when run locally.
	- I ran several tests from the documentations with no issues.
- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING. ✅ 2025-04-28
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `pyproject.toml` file or elsewhere. ✅ 2025-04-28

- [x] Package has a README.md file in the root directory. ✅ 2025-04-28

- [x] The package name ✅ 2025-04-28
    - [x] Continuous integration and test coverage, ✅ 2025-04-28
    - [x] Docs building (if you have a documentation website), ✅ 2025-04-28
    - [x] Python versions supported, ✅ 2025-04-28
    - [x] Current package version (on PyPI / Conda). ✅ 2025-04-28

- [x] Short description of package goals. ✅ 2025-04-28
- [x] Package installation instructions ✅ 2025-04-28
- [x] Any additional setup required to use the package (authentication tokens, etc.) ✅ 2025-04-28
- [x] Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file. ✅ 2025-04-28
    - [x] Brief demonstration of package usage (as it makes sense - links to vignettes could also suffice here if package description is clear) ✅ 2025-04-28
- [x] Link to your documentation website. ✅ 2025-04-28
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem. ✅ 2025-04-28
- [x] Citation information ✅ 2025-04-28

- [x] Package documentation is clear and easy to find and use. ✅ 2025-04-28
- [x] The need for the package is clear ✅ 2025-04-28
- [x] All functions have documentation and associated examples for use ✅ 2025-04-14
- After going through the APIs for all the main functionalities, I am very impressed by the level of details in the documentation and the examples. The authors have clear examples for nearly ever function. Importantly, all calculations are clearly documented.
- [x] The package is easy to install ✅ 2025-04-14

- [x] **Installation:** Installation succeeds as documented. ✅ 2025-04-14
  - I created an issue https://github.com/adtzlr/felupe/issues/953 since the tests werre not working locally for me
	- There does not appear to be any CI for the tests.
- [x] **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide). ✅ 2025-04-28
    - [x] Package supports modern versions of Python and not [End of life versions](https://endoflife.date/python). ✅ 2025-04-28
    - [x] Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass) ✅ 2025-04-28

2 hours
