### 2025-05-16 #212 @adtzlr
Hi, I just wanted to let you know that in the docs every single view on a mesh / field / solid and all matplotlib-plots are (re-) generated during build-time on readthedocs. No static views or plots are used in the docs. All relevant code-blocks in the docs are executed in the build.

@nicoguaro, @crhea93 what I want to say is that in my opinion you don't have to check if the views and plots match the corresponding code-blocks, but that's just a suggestion. I have spent a considerable amount of time in the past that this works out of the box.
