### 2025-05-16 #212 @adtzlr
Okay, thank you. How should I resolve that? I have a Github action for the tests and the same Github action is used to upload the code coverage. This action runs on both the min/max Python versions.

Also, all the doctests are executed while building the Docs on RTD.
