### 2025-05-16 #212 @adtzlr
I could also help out on the **performance** topic: for the examples in the docs, the runtimes, provided by [Sphinx-Gallery](https://sphinx-gallery.github.io/), are shown at the end of each example. On a typical local machine, these examples will run faster compared to the quite low-end readthedocs virtual-machine runs.

The [benchmark](https://felupe.readthedocs.io/en/latest/#performance), which is shown in the docs, depends very much on the machine and the NumPy version used. I've seen best results on a modern Intel CPU with a NumPy MKL-build. On the other hand, my Linux AMD-based workstation with a rather low-end Ryzen 2400G is capable of assembling 70000 dof/s (vs. up to 130000 dof/s on the Intel machine). These differences are only important for computationally cheap linear-elastic materials. For nonlinear hyperelastic materials, much (much!) more time is spent on evaluating the constitutive material response anyway.
