### 2025-09-01 #212 @adtzlr
> [@adtzlr](https://github.com/adtzlr) youve completed everything I've asked for and I've accepted your responses. I believe [@tkoyama010](https://github.com/tkoyama010) is still waiting on another review

Dear PyopenSci team (@tkoyama010 @lwasser), I'm really looking forward to submit my paper to JOSS but I don't think it makes sense right now because this review here is 50% done. What is your opinion on that? Wait until this review is finished or try to submit to JOSS directly? 🙏🏻🙏🏻
