### 2025-09-03 #212 @lwasser
hey there @adtzlr let me check in to see where this review is. It appears that you have one "Accept" from @crhea93. Thank you Carter!!

@nicoguaro are you able to complete the review in the upcoming week? This issue has been open for a year, so it would be nice to move things along.

@adtzlr because you already have a completed review, if you can wait a bit longer, JOSS will accept our review as theirs. Otherwise, you will have to start over with JOSS. Let's see if we can move things along here to make life easier for JOSS. They face the same editor/reviewer challenges that we do, in that people are busy and a lot is happening in the world right now.

Please do me a favor and ping me here if we need to move things along again.
