### 2025-09-17 #212 @nicoguaro
@adtzlr, I have finished the second round of my review. Since I saw so much activity from your side during the last year, I just downloaded and installed the latest version to check it out. I think that the package is turning into something really useful. Congratulations.

I do have some tests failing on my computer. This is the output of pytest

```bash
=========================================================================================== short test summary info ============================================================================================
FAILED test_constitution.py::test_umat_hyperelastic - _tkinter.TclError: Can't find a usable init.tcl in the following directories:
FAILED test_field.py::test_axi - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED test_field.py::test_toplevel_merge - AttributeError: 'FieldContainer' object has no attribute 'merge'
FAILED test_math.py::test_math - AttributeError: module 'felupe.math' has no attribute 'revolve_points'
FAILED test_mechanics.py::test_load - TypeError: newtonrhapson() got an unexpected keyword argument 'callback'
FAILED test_mechanics.py::test_checkpoint - AttributeError: 'SolidBody' object has no attribute 'checkpoint'
FAILED test_mechanics.py::test_checkpoint_incompressible - AttributeError: 'SolidBodyNearlyIncompressible' object has no attribute 'checkpoint'
FAILED test_mechanics.py::test_axi_to_3d - AttributeError: 'SolidBody' object has no attribute 'revolve'
FAILED test_mechanics.py::test_axi_to_3d_mixed - AttributeError: 'SolidBody' object has no attribute 'revolve'
FAILED test_mechanics.py::test_axi_to_3d_incompressible - AttributeError: 'SolidBodyNearlyIncompressible' object has no attribute 'revolve'
FAILED test_mechanics.py::test_axi_to_3d_quadratic - AttributeError: 'SolidBody' object has no attribute 'revolve'
FAILED test_mesh.py::test_read - meshio._exceptions.ReadError: File tests\mesh.bdf not found.
FAILED test_mesh.py::test_read_nocells - meshio._exceptions.ReadError: File tests\mesh_no-cells.bdf not found.
FAILED test_mesh.py::test_expand_revolve_quadratic - KeyError: 'quad9'
================================================================================= 14 failed, 156 passed, 11 warnings in 55.29s =================================================================================
````
