### 2025-09-17 #212 @adtzlr
Hey @nicoguaro ,

could you please describe the steps you have done on installing FElupe? The errors are all related to new features, so it seems that you've combined new tests with an older version.

CD/CI pipeline is successful on all versions and the main branch, that should be resolved quite easily.

Could you also please try to run your local tests in a fresh environment?

Thanks :pray:
