### 2025-09-17 #212 @nicoguaro
@adtzlr, I installed FElupe using pip. Nevertheless, I ran the test in the source code in the main branch on GitHub. I did run the tests in a fresh environment created for the sole purpose of the review.
