### 2025-09-17 #212 @adtzlr
Okay, thanks! I think this happens in all Python packages which have the source code located in /src. In the test files, FElupe is imported as `import felupe as fem`, i.e. from the pip-installed version and not from the locally cloned copy of the main branch. Please don't expect that newer tests will work with older versions.

If you use a new virtual environment, either install the development version or the latest release from pypi. Then, use the tests for the installed version.

FElupe is a pure Python package, that means if you simply add its /src directory to your PYTHONPATH environment variable, it will work without installing it (if you have all dependencies installed manually). But this is not the recommended way of using FElupe. The editable install as shown in the docs is a much better alternative.
