### 2025-09-17 #212 @nicoguaro
I understand. I tried installing with pip and the local version and it doesn't work. Check https://github.com/adtzlr/felupe/issues/991.
