### 2025-09-18 #212 @adtzlr
> I understand. I tried installing with pip and the local version and it doesn't work. Check [adtzlr/felupe#991](https://github.com/adtzlr/felupe/issues/991).

Thanks for pointing that out. It is now resolved, see https://github.com/adtzlr/felupe/pull/992.
