### 2025-09-18 #212 @nicoguaro
@adtzlr, that did the trick. I have just double-checked, and it is now working.

@lwasser, from my end, this package is ready for publication. Regarding acceptance in JOSS, I think that the paper is missing the references, and that's it.
