### 2025-09-18 #212 @adtzlr
> @adtzlr, that did the trick. I have just double-checked, and it is now working.

Thank you for your review @nicoguaro!

> @lwasser, from my end, this package is ready for publication. Regarding acceptance in JOSS, I think that the paper is missing the references, and that's it.

The references are inserted into the final pdf file. I did not include the draft pdf in the /paper directory of FElupe yet. A draft preview with all the references is available here: https://github.com/adtzlr/felupe-paper/blob/main/paper/paper.pdf
