### 2025-09-26 #212 @adtzlr
Hey @lwasser, @cmarmo, @tkoyama010,

now that the reviews are done, what are the next steps? Anything required from my side? Thank you 🙏🏻. And also please tell me once its possible for me to submit to JOSS (I have to do the submission by myself, right?).
