### 2025-09-27 #212 @cmarmo
It looks like Felupe has been approved by pyOpenSci! Congratulations!
Many thanks to @crhea93 and @nicoguaro  for reviewing this package!

@tkoyama010  below the author and editor wrap up tasks
I' m going to check some of them myself, do you mind follow-up in the next days?
@adtzlr thank you for you patience! Please have a look at the author wrap up tasks below, and yes feel free to submit to JOSS, referencing this issue, letting the JOSS editors know that you already have your package accepted by pyOpenSci.

- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

---
