### 2025-09-29 #212 @adtzlr
Dear @cmarmo, @tkoyama010,

I added the badge to the repo, now included in the latest FElupe release v9.4.1. I filled out your post-review survey and the submission form for JOSS. I can of course write a blog post on FElupe, but please give me some time (at least two weeks).

Thank you!
