### 2025-09-30 #212 @tkoyama010
> Invite the package maintainer(s) and both reviewers to slack if they wish to join.

@nicoguaro @adtzlr Are you interested to join our slack?
