### 2025-10-06 #212 @nicoguaro
> [@nicoguaro](https://github.com/nicoguaro) Are you interested to join our slack?

Yes, please
