### 2025-10-07 #212 @lwasser
Friends - if anyone wishes to join our slack - you can email me at leah @ pyopensci.org or @tkoyama010 can invite you as well if  he has your email!! @cmarmo thank you ❤️ and thank you everyone for supporting and participating in this review 🚀
