### 2025-10-13 #212 @tkoyama010
@adtzlr @nicoguaro Please join our slack from [here](https://pyopensci.slack.com/join/shared_invite/zt-39qitgkqb-gZTIo79xCJhS5kSxW1yNfg#/shared-invite/email).
