### 2025-11-17 #212 @lwasser
Ok, everyone - I think we just need the JOSS DOI to wrap up and close this review. Can someone please provide it? @tkoyama010, then you can add it to the top issue comment (the YAML there) and close this issue as completed.

Thank you everyone!
