### 2025-11-17 #212 @adtzlr
Here's the doi: 10.21105/joss.09160
